[roll] Roll fuchsia [netstack3] Deflake routes inspect test
We're getting empty inspect data because this test is missing
synchronization with the out dir actually being served.

It's a better practice to read the loopback interface's ID either way so
we do that and deflake the test.

Original-Fixed: 132923
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917305
Original-Revision: e887ce1ee4960cbc1094f1b75f279db32c1c1490
GitOrigin-RevId: 5b9dfbbce07e746fec17330c6b6dabddfd6d43ca
Change-Id: Ia5411ef9c301ba24023464756ad78ffe4c9b434b