[roll] Roll fuchsia [lowpan-ot-driver] Make sure meshcop state bitmap reflects agent state
Previously, we were always setting the `CONNECTION_MODE` on the state
bitmap to `CONNECTION_MODE_PSKC`, regardless if the border agent was
actually accepting commands or not.

This change makes sure that the state bitmap accurately reflects the
state of the border agent.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/916364
Original-Revision: d0693fcfbee1298766fe5d3d9a4148519d9409fc
GitOrigin-RevId: 7259ff91dddad33cb4c2d1162b0da469a18a3cbe
Change-Id: I9e390937b9214112fffc7d3830920be6e23712d4