[roll] Roll fuchsia [starnix] Add Perfetto consumer
This perfetto consumer communicates with a Perfetto daemon running
inside the Starnix container via its unix domain socket interface.
When Fuchsia tracing starts, the consumer indicates to the Perfetto
daemon to start tracing inside the container as well, and when Fuchsia
tracing ends it collects the data and writes it to the Fuchsia trace.

Some additional files are added to the tracefs implementation so that
the Perfetto daemon will believe it is running with a functional
tracefs, which allows us to use its ability to control ATrace on images
that use it.

Tested: See tqr/760470
BUG: b/291094354
BUG: b/271467732
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891188
Original-Revision: 963fdc7f0a7946dd46613dc13c3f7003d6ee7656
GitOrigin-RevId: 5e34d3224d8bb98379ff81b26ada911e586ee9d7
Change-Id: Ie1ec84051e77d418e43c5f3223e4fb6f752fe55e