[roll] Roll fuchsia [Starnix] Keep zombies in the PID table
Previously zombie processes were not kept in the PID table. As result
this kill() was failing for zombies, which is unexpected. Now a
reference to ZombieProcess is stored in PID table. This allows to
fix kill() to handle this case correctly. Also in the future it will
allow to show zombie process status in /proc

This fixes flake in sigwait01 test, where kill() was racing with
process exiting.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907711
Original-Revision: edc968e84c71c1f7d1cef0693295c7228f112cb6
GitOrigin-RevId: 5c81254460e99b3e0c31c663dca10c0f29df7dbc
Change-Id: I93adcc4fd984e3b56b7f873a983f01efeaaa9a77