[roll] Roll fuchsia [driver][sdk] Migrate to DriverRegistration
This change migrates the driver component libraries to use the FIDL
based DriverRegistration mechanism to export the driver. This involves
moving away from the legacy C lifecycle ABI.

There are also some changes needed in the compat driver as it was
using the CustomFactory internal feature which is not available
anymore. Instead the a server has been created just for the compat
driver so it can do the custom initialization logic it needed.

In order to not break existing drivers and force changes for them to
recompile, all of the customer facing APIs have been unchanged. Instead
the internal machinery has been changed to work with the new mechanism.
This includes the DriverBase methods users should overwrite and have
access to, the DriverBase constructor they should use, the testing
library APIs, the export macro usage, the async completers for start and
preparestop, etc...

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891795
Original-Revision: 3e0f1c2f8e143ded7c25180bde66e75b4da4e18a
GitOrigin-RevId: 9d77ac366ada7c07c74fc056096d502db147d7b6
Change-Id: Icc45729762b278bc7402d5bc1f1026e13f1e5a88