[roll] Roll fuchsia [assembly] Allow de-duping for config-data
This can be turned on using a constant at the top of
assembly_builder.rs: DEDUP_CONFIG_DATA.

Allow the same config-data file to be added in both platform_configuration
and the legacy bundle as long as the contents are identical.

This allows us to carefully migrate config-data files from the legacy
bundle into platform_configuration.

We do this by replacing the usage of NamedMap with NamedFileMap which
gives us access to the function 'add_entry_as_blob' which finds the
merkle of the file and de-dupes as necessary.

Original-Bug: b/300329753
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917273
Original-Revision: 7a2a7552ea284fc9f2c4da143755552484554e36
GitOrigin-RevId: f47f72bde49423d02378a637bc905ba80a3dff38
Change-Id: Ib0c33faf09b9d9e4e4c8effd3871c8f573de9e65