[roll] Roll fuchsia [ffx][overnet]Adding ABI check to remote_control_runner
This updates the runner script on the target to check a provided ABI
revision to the one used to compile the runner.

This result is then returned in a JSON object before starting the
Overnet communication.

Support is also included to fall back to invoking the legacy
runner if the abi revision flag is not known.

This change updates the FIDL library to include an enum for the status
of the compatibility check, and then exposes them via the Serializable
CompatibilityInfo struct.

Original-Bug: b/295899889
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/916357
Original-Revision: 951b8e582c675407fa2cb46c775ec89237400258
GitOrigin-RevId: 39a925e66db3a5842d3a440adc71d192942d4c93
Change-Id: Ib867a3e75f55221b56ccfd917ad805d7c2c4965b