[roll] Roll fuchsia [bundles] Add placeholder bundle for tools used by shac
Some static analysis checks run by shac depend on tools produced by the
build, so those tools must be available to shac when it runs in the
static-checks builder. We can do this by providing a bundle for the
static-checks builder to build, which contains all the tools that shac
depends on.

Original-Bug: 82386
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917112
Original-Revision: 2b5221e03351afc58a81e83097274777d83d614d
GitOrigin-RevId: 5e47e099f68d6bbc8ac518a064dfae1ad3a73de6
Change-Id: Ibdc1f7778f385764ccf828bbd1a60038c1b54c7d