[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc][GPU] Fix dependencies for externally installed stub files (#66653)
Summary:
The GPU build has a lot of magic around how we package the output.
Generally, the GPU needs to exist as a secondary fatbinary image for
offloading languages. This is because offloading languages pretend like
offloading to an accelerator is a single file. This then needs to be put
into a single file to make it mesh with the existing build
infrastructure. To work with this, the `libc` makes an installed version
of the library that simply embeds the GPU code into an empty stub file.

This wasn't being updated correctly, which lead to the installed `libc`
static library not being updated correctly when the underlying file was
changed. The previous behaviour only updated when the entrypoint itself
was modified, but not any of its headers. By adding a dependcy on the
actual *object* file we should now capture the regular CMake semantics.

GitOrigin-RevId: d3e04ce5c62fee14d4504db3d9ae7ab105fb303c
Original-Revision: 1ec4d4c31102e480633405056d703bbb9ffb2e72
Roller-URL: https://ci.chromium.org/b/8769592476196418721
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918214
Original-Revision: a0ac67d21eaa42db8c35c4619e6b3c6075de4218
Change-Id: Id4f854413242fecc9298b6609840ab2fca53f305