[roll] Roll fuchsia [llvm-profdata] Avoid coverage instrumentation for UEFI
The llvm-profdata runtime can't be compiled with instrumentation
for the UEFI targets.  This was already worked around for the
backend profile instrumentation, but the same workaround is
needed for the coverage instrumentation too after recent upstream
changes in the instrumentation passes.

Original-Fixed: 133821
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918252
Original-Revision: c0cfeeb0ebc9181d5885ea13effbfdeabc89b9e1
GitOrigin-RevId: 329a3413b70e7b63ab470bb606b0ac0a9bb3fd87
Change-Id: I9491171dc19246f621a3f6c745c28551bd2f00b4