[roll] Roll fuchsia [vsock] Fix CreditInfo counting
Fix the credit values in CreditInfo.
Originally, we were setting fwd_cnt to how many bytes were in the
socket, but fwd_count is supposed to be the number of bytes consumed.

The incorrect fwd_count would lead to deadlocks because when the bytes
received equalled the buffers size we would report fwd_cnt = 0. This
meant the device thought the buffer was full when it was actually empty.

From the spec, for calculating sizes:
    /* tx_cnt is the sender's free-running bytes transmitted counter */
    u32 peer_free = peer_buf_alloc - (tx_cnt - peer_fwd_cnt);

This is now unit tested. The internal classes of SocketDevice are made
public for testing.

Original-Bug: b/300123548
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918296
Original-Revision: de828153cb4fcd30036a35b206c54d45100686d1
GitOrigin-RevId: 63b37dab9b10a1a2a2dba445a06b22ee70907efb
Change-Id: I79b925ff2dfcd039689b7623e6aef5d062836c53