[roll] Roll fuchsia [driver][sdk] Fix breaking Stop in test lib
In the migration cl to the driver registration, I made a breaking change
that changed the definition for the Stop call in the test library's
DriverLifecycle class. This reverts that breaking change. The
PrepareStop hook remains the entrypoint to the FIDL Stop to keep
consistent with the naming in DriverBase.

Original-Bug: 133230
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918335
Original-Revision: 9ef32e3401115294d232396cf3ded9e1d90128c7
GitOrigin-RevId: 77731ad099d7437ae0498eb0c4263b8d53a6aaa8
Change-Id: I679ed8d985942554b00e31885ffcbdb8eefadece