[roll] Roll fuchsia [driver_runtime] Fix timer cancel bug.
Switch from async::Task to async_task_t.

async::Task is expected to run in a single-threaded environment,
so will assert if cancelation fails.

It's not a concern if cancelation fails, that means the timer just
woke up on another thread, and the handler is about to run.

Original-Bug: 132992
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/913893
Original-Revision: dda512e8622739797c7ed0f7cf77d586ffe1107c
GitOrigin-RevId: 8c3c0e59ad657babd968e63eee1fd0a870bded09
Change-Id: I8c352076949c02df75e8c4cc46b799d8464a64cf