[roll] Roll fuchsia [starnix] Fix race on the binder driver
The previous implementation was releasing the same object multiple time
when an object had both strong and weak reference dropping to 0 in the
same cycle. That creates a race when the same object is added back
between the 2 removes and the new object, that is still references, is
then removed by the second removal.

This CL fixes this race by not removing the object immediatey. Instead
if keep track of all object that could be removed, and remove them at
the end a single time.

Original-Fixed: b/298935909 b/291922706
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917973
Original-Revision: 4434e4e4ad9076216807cc1798e629bad189a1b8
GitOrigin-RevId: bfcb0f958c89ecb2d5f53a717d26b0f3f2beeafc
Change-Id: I753eee4ff6c38c149bea6d9bd99fdbe812a343bd