[roll] Roll fuchsia [vsock] Fix rx-credit deadlock
Fix a deadlock that can occur when the driver receives too much data.

During periods of high amounts of data coming in, the device will not be
receiving any credit update messages for it to know that the socket
buffer is being drained. The solution to this in this CL is very
imperfect in that we just send a credit update after processing an
incoming RW message. This does not guarantee deadlock won't occur. We
should instead be using eventing off the `ZX_SOCKET_WRITE_THRESHOLD` so
that we can send credit updates only once the socket buffer drops below
some low-water mark instead.

Original-Bug: b/300123548
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917555
Original-Revision: 56e2341e1e96b40be69594c5cb4fa007f9313250
GitOrigin-RevId: c95b476521834bdf563bb6ed7a6744e5acf61f93
Change-Id: Ie52e86626b7b73e37e4dd8724f809fc8ca1d2a04