[roll] Roll fuchsia [fdf] Fix uaf in driver_server start
This use-after-free was occurring when the completion of the
StartCompleter was running before the dispatching of the Start from the
WireServerDispatcher (internal FIDL machinery) had the chance to fully
run. At the end of the dispatch handler, the request which lives in a
RAII type fidl::DecodedValue is released and the destructor is run. By
the time that this was running, in the error case, the arena had already
been freed up at the end of StartCompleter lambda, and the underlying
data (which the DecodedValue was wrapping) was already gone. This led to
a use-after-free.

The fix is to post a task to run the Start so that the callback is
guaranteed to run asynchronously and allow the cleanup on the
WireServerDispatcher to take place first.

Original-Fixed: 133776
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918178
Original-Revision: 1ab8d7ad194bbf0eab72264a38378a273f800099
GitOrigin-RevId: 9b4cd95ab49b190694eaf96cab42158af487211c
Change-Id: I0e98e191356941a85e83d0e511c6bd8013b425c5