[roll] Roll fuchsia [brcmfmac] Forward boot metadata
Add boot metadata to the `wifi` composite device on astro, nelson, and
sherlock devices then forward that metadata to the `brcmfmac-wlanphy`
device. The `brcmfmac-wlanphy` device needs access to that metadata and
will not get it when DFv2 is enabled unless that metadata is forwarded
to it.

Original-Bug: 133538
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/916392
Original-Revision: b897be37458f1d482dd9e62107dae935a937181d
GitOrigin-RevId: ee6814a48c9fe3666f34b56bd671c45ba81f05da
Change-Id: Id3794e1763c0df3bee7d0e17c34dcc2fbd3aac24