[roll] Roll fuchsia [fidl][examples] Adding echo-realm target.
The build file in //examples/fidl/echo-realm does not have
a target that matches the directory name. This causes confusion
when trying to refer to it in a build.

This adds a group named "echo-realm" which depends on the "echo_realm"
target.

Fixes: b/297069811
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/906818
Original-Revision: 0c99d1afe4f3ae4c0d9df783546aeaf6d8f53ccc
GitOrigin-RevId: 40062792262c435538df71205d8a1275b14e2bc3
Change-Id: I789ffe6580d262b6fd374288fa048de82c16250d