[roll] Roll antlion Use auth/assoc hostapd checks for WNM tests.
A recent change introduced the ability to check the state of a DUT
on the AP side (whether it is authenticated, associated, etc.).
This change updates the WNM test to use these checks. A few minor
updates to the tests are also included:

  - correct the operating class for 5 GHz VHT AP
  - rewording of some assert failure messages

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/antlion/+/918194
Original-Revision: 862a5d82470c2b7967cdbc62f88208ec1b154aa0
GitOrigin-RevId: 93bdaa41ecd03b7234db9e7043aca254ce8d4a88
Change-Id: I31e5cb5a38150187f975c1639ca0a7592ee057eb