[roll] Roll fuchsia [sys][core] Deduplicate debug-dash-launcher
This component is declared in two places, remove the
declaration in core.cml so it is only declared in a
core shard.

This allows us to move the offer routes into the core
shard.

This also allows us to optionally define the offer
capability up to the root CML which allows us to
remove the security exception.

Original-Fixed: b/295357307
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907362
Original-Revision: cb98f08802c40518b133ee5e0aa396897cfe5d3b
GitOrigin-RevId: af4629d0874f31c20bedde6852bc4a03d660d4ad
Change-Id: I9435b38d704981ec9b9fe924cabc1201dd19979c