[roll] Roll fuchsia [ffx] Remove automatic connection to targets
Instead of connecting a host-pipe to a target on discovery, just note
the target exists, and connect on demand  (i.e. after a request to
the daemon such as `OpenTarget` or `OpenRemoteControl`). Behavior is
basically transparent, except that undiscovered targets will remain
unmolested, and the initial connection will be slightly slower.

(When the daemon starts, it will connect to manual targets in order
to learn their nodenames. Eventually, we will change the daemon to
store the nodename when it is first added -- b/300986641.)

Original-Bug: b/297896623
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/908842
Original-Revision: 5c4357d8de7ddd9eb6e043ba40acd0b8327af148
GitOrigin-RevId: fcb3282a8fd52fd7688f19ca9f08f4a854853e17
Change-Id: I047f748918fb28ffbe3e5bd59d8949ba4b18ada0