[roll] Roll fuchsia [fidl][rust] Allow source-breaking opt-in
Rename the __Unknown flexible enum and union variant so it clearly
reflects that it is source breaking. Remove the non_exhaustive
prohibition and deprecated nudge.

For tables, rename the field to __source_breaking and type it to
the newly introduced fidl::marker::SourceBreaking marker.

That allows users to willingly opt-in for source-breakage, which is
particularly useful for in-tree code when used with care and
responsibility.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917715
Original-Revision: 9ba4bc0b95baeec760145c0f1f48a5a6085552a4
GitOrigin-RevId: 23530d4290fc6a241b535c616e898d103bdbba6c
Change-Id: I3c5f1a3cade20c92da4b414b7a3a9a4c0c2db19f