[roll] Roll fuchsia [connectivity] hw-sim should include realm_builder_server
The hw-sim tests were trying to use realm-builder, but it wasn't
included in the package, which caused it to fail in infra (although it
doesn't appear to be a fatal error, since the tests passed).

This adds realm-builder as a subpackage dependency, since that's the new
way we expect this package to be used.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919259
Original-Revision: af89163ec5bfbf7da5a197d528615b4f34043f44
GitOrigin-RevId: 93c17997933a0ecb76b0e1e69f7eeffc6a95a5b3
Change-Id: Idb6acda44a756167986c03b7967e29e8efe81bc8