[roll] Roll third_party/pigweed pw_libc: Pull in 'abort'
abort is platform specific and is found in baremetal/,
generally we would prefer making a seperate source set
which uses stdlib/baremetal as the dir and pull in abort
from the `functions` list. Here it isn't necessary because
it is the only function we need from baremetal/ and it has
no associated test anyway.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/138518
Original-Revision: 7970ac98c81a9a00dfba6242d0b607b633fc2522
GitOrigin-RevId: 62ddb8b1934dbca3f07480f0e364c5bdee48dbb0
Change-Id: Ia42782a8575d01cf941bf27f8e24652d23837a85