[roll] Roll fuchsia [netstack3] Implement ICMP sockets in terms of datagram module
In this change, `connect`, `bind` and `send_conn` are implemented for
the ICMP sockets in Core. The tests are testing the same behavior they
used to, except for sending or receiving, the old behavior is that the
buffer only represents the body of the ICMP echo, the new behavior is
that the buffer is the entire ICMP echo packet, including the ICMP
header.

Original-Bug: 125482

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917534
Original-Revision: d961276c275f2e4b4ff1ae0bf08e057bf9a28f91
GitOrigin-RevId: 03cb213dbc8bcd6244c0f41e5d18ea6c3bd97caf
Change-Id: Icf6d26f537a42234919dd225464c2e15de082550