[roll] Roll fuchsia [shac] Integrate with doc-checker
Previously doc-checker integration was duplicated in `fx format-code`
and in the static-checks infrastructure recipe, but now it's
consolidated in shac.

We still continue to run it from `fx format-code` even though it's not a
formatter to preserve the behavior, but I'm hoping we can remove that at
some point.

Original-Bug: 82386
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919555
Original-Revision: e75ed20a38785fe15c0da89b33635885e271c09a
GitOrigin-RevId: 01a52d2b4adab47ec7de612eb4945b025ac3c6c7
Change-Id: Ifc1813048012b90db32543c55d77310b6ec7ab79