[roll] Roll infra/recipes [checkout] Add commit step to submodules for patches.json
When patches.json changes are applied in checkout, if change is applied
to a submodule, fuchsia.git jiri update will override the revision
pinned. We therefore need to add an extra step to commit the revision so
that submodules revisions are kept. Here, we need to compare projects
stated in the json file with jiri projects' name and find which projects
are labeled as submodules.

Original-Bug: b/300977004
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/919613
Original-Revision: 5010317aa61cd9a3e4b5989890853e98fbd672a8
GitOrigin-RevId: a26a23b5bdf4ef0b5e35989ee4f160d69b626e39
Change-Id: Ia7328b9a5ed3178dd65968ff3214401d1fca1f7c