[roll] Roll fuchsia [fidl] Include external flexible method result structs
Flexible two-way methods that don't have an error type specified are
actually implemented on the wire as a struct with a union that has a
success and a transport error member.

For composed protocols the definition of that top-level response struct
must be included in the external_struct_declarations IR field. It was
for methods with errors, but not for flexible methods.

Test: Added test-cases to the fidlc testdata, verified that they broke,
then verified that they're fixed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/915439
Original-Revision: c1fcfe86c94458326c0a650c6879c2e6b134be50
GitOrigin-RevId: d911725c4470af1c911080f895d8492a6697a625
Change-Id: I4c10dfc39256ab326e7fdcc40e925dce53e0c22d