[roll] Roll infra/recipes [mirror_version] Add recipe to mirror versions across two repos
The "source" repository contains the release versions. The "destination"
repository is where the release versions should be mirrored.

The intention of this recipe is that LUCI scheduler will fire on
individual commits on a per-branch basis, resulting in this recipe
checking each commit for its "origin" (source) commit and mapping that
back to its release version, if there is a release version.

If there is, then push that same release version to the destination
repository.

Original-Bug: 132016
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/920113
Original-Revision: f5b06c5cf12ba7d2479694434ef83504033c1c22
GitOrigin-RevId: d74b8e030de916527a7d7dccf165f6b00227a5aa
Change-Id: I325b2260fb75e9b08d257ae73ce5f9e80290ec07