[roll] Roll fuchsia [fuchsia-controller] Fix result type import bug.
If there's a FIDL library that is being imported
and there's a struct that contains a channel to a composed protocol, and
the composed protocol has a method that returns a result,
fuchsia-controller would throw an exception. This fixes that and
includes a test reproducing this issue in test fidl code.

Original-Bug: b/301178571

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919196
Original-Revision: 94e912fe78eb00eb2385f7d77a293bef687bc96b
GitOrigin-RevId: 67268330bcd1e78c00fbd01a0d6873fadd727503
Change-Id: I388a142cbb4722f6c448ab62f9af6e0d1980d0a7