[roll] Roll fuchsia [fxtest] Use separate script for matching
The matching we do is expensive, and the version of jellyfish we use for
approximate matching implements the algorithms in Python itself.

This change instead depends on the new dldist program for approximate
matching, each string is passed on a separate line and the matching
output lines are the distance measurement for that input line as a
decimal integer.

This change does the following:
- Call dldist for matching decisions.
- Calls to dldist are added as children of "Matching" events, matching
  is now asynchronous.
- An option is added to the event group payloads to indicate that status
  updates should skip outputting the (verbose list of) children for each
  matching operation.
- Console output is changed to respect the hide children option.
- dldist has an option to match the prefix logic that was previously
  implemented in Python.

Original-Fixed: b/300157465
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918217
Original-Revision: ed9f44be8798d9e0e7a820510e958690aee3353e
GitOrigin-RevId: 9d0bc68851a06b897a4209fce0d3429d64c53b8e
Change-Id: I42be84071980f3a208d0796c4e5d18f22a77310f