[roll] Roll fuchsia [fxtest] Do not suppress color output of fx build
fx build detects if its output is not to a TTY and suppresses color
output. This change uses the appropriate env var to force color output.

This required changing the semantics of the env argument to AsyncCommand
to be to augment the incoming environment, which may require further
tweaking in the future if removal from the environment is needed.

Original-Fixed: b/299145560
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918188
Original-Revision: 266176ea00363cefa4d37fd653dbbb6bb5a0c699
GitOrigin-RevId: 5fec722917b12a2b9401e504954283a9e1a96bc4
Change-Id: Icbe20a0e37630e66611748e253c27f017e7365eb