[roll] Roll fuchsia [fidl] Put @transitional behind an allowlist
The @transitional attribute was added many years ago to simplify making
FIDL changes in the layer-cake era. We now have a more authentic
monorepo and @available to support out-of-tree changes.

It's easy to build incorrect software when using @transitional and we've
seen @transitional left on methods for more than four years. It's time
to transition to our shiny new glorious versioning world.

This adds an allow-list for the use of @transitional so that as we
eliminate legacy uses of it no more are added. Right now that's behind
the experimental flag transitional_allow_list because there are FIDL
files out of tree that use @transitional, but once those are gone this
should become default behavior.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919714
Original-Revision: f1989d9242bda6ab1f79788f95fd26a5adceb382
GitOrigin-RevId: b97462d3a709d53a45fef04a382ec8ea079d5f32
Change-Id: I0b483905d11e85bedf4f4dbb4cb458a9ea012452