[roll] Roll fuchsia [paver] move error logs to actual errors
ExtractZbiPayload() is used to check whether a blob is a ZBI or not, so
we shouldn't be logging errors when it's not; this is an expected use
case.

Error logs are still printed if a blob does look like a ZBI but is
malformed. Also adds one error log to IsValidKernelZbi(), which does
always expect to see a valid ZBI image so wants to log if the blob
doesn't look like a ZBI at all.

Original-Bug: b/300655645
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921135
Original-Revision: a56941bace641ebca4f3e9c0ae65b9df87bc031b
GitOrigin-RevId: 6145276e1659d3abf56dce2e417b0534f43c4273
Change-Id: I47bac288d7d8150f4ebe0ecf9eb88de2231c07af