[roll] Roll llvm-project/libc [libc] Fix cyclical dependency on errno matcher for NVPTX architectures
Summary:
The NVPTX backend cannot handle cyclical dependencies on global variable
initializers. That is, a global variable cannot be used to initialize or
reference another global variable inside of it. This situation was
encountered with the new errno tests. This patch simply replaces the
offending function with a constant version to break the dependency and
alllow the tests to run again.
GitOrigin-RevId: 889bfebe850c7bf879680a28d3e5b7cdf19096b9
Original-Revision: 930f2cc0c14d5fe46f9c2306f4adeac7d94c9d82
Change-Id: I168c31dbaee494113ac331e758c09d5371e0c6ae