[roll] Roll fuchsia [utest][resource] Fix resource tests on vim3
The resource tests start by scanning the entire address space for a
valid MMIO region to test against. Furthermore, this probe test asserts
that the only failure we can encounter when creating an MMIO resource is
a ZX_ERR_NOT_FOUND.

This approach works so long as the entire resource range being scanned
is accessible to the root resource (see
https://cs.opensource.google/fuchsia/fuchsia/+/main:zircon/kernel/object/resource.cc;l=47
for more information on when this is not the case). This is not true on
Vim3s, where the first 64K is allocated to be free RAM, and is therefore
not accessible via the root resource. This was causing the
`ProbeAddressSpace` test to fail with a `ZX_ERR_ACCESS_DENIED`.

This change fixes the issue by allowing the return code of a failed
resource creation to be either `ZX_ERR_NOT_FOUND` or
`ZX_ERR_ACCESS_DENIED`.

This change also fixes the `InvalidArgs` test, which was failing the
invalid size case because we were once again passing a base address
(1024) that was inaccessible to the root resource. Naively, one would
think passing `mmio_test_base` to this function would fix the issue, but
it causes failures on the other platforms because `[0x0, UINT64_MAX]` is
a valid address range, but fails the `validate_ranged_resource` check
because there are obviously non-peripheral ranges in that range. So,
instead, we keep passing 1024 as the base, but add the same integer
overflow check that would trigger the ZX_ERR_INVALID_ARGS originally
(https://cs.opensource.google/fuchsia/fuchsia/+/main:zircon/system/ulib/region-alloc/region-alloc.cc;l=337)
at the beginning of the `zx_resource_create` syscall, before we try to
run a validation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921136
Original-Revision: b60ca070aeaeb0b7fb3ac95cfdfa3a8562035912
GitOrigin-RevId: 1a5531d47a8ec503538ceca5b4172581b16bc3a1
Change-Id: I0cefc25ab3f4a309a5215d6c3790531f8011dc22