[roll] Roll fuchsia [cmc] Improve dep tracking for capability sources
Some capability types (`storage`, and soon `dictionary`) themselves have
a source in `from`. For example, storage depends on a backing dir, and
dictionaries can depend on another dictionary which they're initialized
from.

The dependency tracking strategy for this today involves replacing the
dependency on the capability with a dependency on its source. This may lead to
spurious cycles. Instead of replacing the capability with its source
dependency, represent the capability as a node in the dependency graph,
and add an edge between it and the source.

As a bonus, this should make the error more useful.

Original-Bug: b/300500098
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921041
Original-Revision: 13b2aa72e0cfef263e5e59f8717a6961980cb342
GitOrigin-RevId: aa16bdf704639cfde313f385af37d81783cd8abe
Change-Id: Iaa26eeaf0dc5999d3eefdd1bf5c0cfb944077472