[roll] Roll fuchsia [ffx][overnet]Adding ABI check to overnet
This updates the runner script on the target to check a provided ABI
revision to the one used to compile the runner.

This result is then returned in a JSON object before starting the
Overnet communication.

Support is also included to fall back to invoking the legacy
runner if the abi revision flag is not known.

The result of the comparison is stored in the target object
so it can be used to inform the user of the compatibility status
and protocol versioning decisions can be made.

Original-Bug: b/295899889

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/901721
Original-Revision: 3f169eaf010bedb60f3439956687441acc6ec917
GitOrigin-RevId: 89d123613ac3239564006809d492d05639d7daad
Change-Id: I91b6dea5dfd8cf9acd64988998120d57fe192e45