[roll] Roll fuchsia [kernel][pager] Fix InvalidPagerSupplyPages on Vim3
The `InvalidPagerSupplyPages` test was failing on Vim3 because it tried
to create a physical VMO at address 0x0, which on a Vim3 is not within
the range of peripheral memory. This CL fixes the issue by using
`vmo_test::GetTestPhysVmo` to create the physical VMO in a valid region.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921539
Original-Revision: 2fefad51a2863a7607c158ee636c51bc09346385
GitOrigin-RevId: 633e43b24c07b309e7b23fd30615969c7543efa2
Change-Id: I7cad04439c49351a738648ae969ab7b5d5ae1cbe