[roll] Roll fuchsia [fuchsia-crate] Ensure that the main() Result isn't logged twice
The previous state of this macro would log the Result::Err to syslog,
but then return the same result which would trigger
Termination/report on the result, which would print the Result::Err
to stderr too.

This fixes this problem and guarantees that we only log to syslog.

Original-Fixed: b/300155961
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919913
Original-Revision: 43553a0dfd91d50e93977a8b9be3bd72361345de
GitOrigin-RevId: e90d70c5187e87026205dabe16778ef214fbbdcc
Change-Id: I49dceced4c6e813e11ad3a550d42d68b9406a38b