[roll] Roll fuchsia [sys/component_manager] support use runner
This CL adds the ability to 'use' runners in CML files. Runners may be
used from the component's parent, a child, or from the framework. FIDL
and internal Rust representations of CML documents can also use runners
'from environment', but this source is not allowed in CML files nor the
Rust representation used for deserialization.

Specifying a runner in the 'program' section is still supported.
Contrary to the design specification for this feature, the translation
from the internal Rust representation to FIDL (and therefore CM) of
manifests that contain program/runner does not add a 'use from
environment' entry. This feature currently provides no tangible value
and invalidates test goldens in dependent repos. This change can be
made in a separate set of CLs, if desired.

Original-Bug: b/296274455

Test: new and updated unit tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/915433
Original-Revision: ef09fc1cc70916d12af04b000a3ebbdc0a468844
GitOrigin-RevId: 0ca813e0f5e0415fe0a9ee69dd67179e73d3d315
Change-Id: I93dc6e445ae3ca3d0f741f452d96142cd2a652c6