[roll] Roll fuchsia [starnix] Keep mount information in overlayfs
The mount information is required because if the upper directory was
mounted read-only, it should not be possible to write to it through a
overlayfs mount.

This CL doesn't yet fix this issue, but it lays the ground to allow it
once access to the node will requires the mount information.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919239
Original-Revision: 885495f1ca601654840f6d50ae99fb6bf75121e2
GitOrigin-RevId: 92e817630796392076a6c0b8d11e5cc6167afac3
Change-Id: I599b37418497ff3cb64bbd07f0299a896abb625e