[roll] Roll fuchsia [starnix] Implement MADV_DONTFORK and MADV_WIPEONFORK for madvise()
This implements the DONTFORK and WIPEONFORK advice values for the
madvise() system call. DONTFORK marks a mapping such that it will not
create any mapping in a cloned address space. It can be undone by the
DOFORK advice. This advice follows a memory mapping through remapping
and extension via remap(). WIPEONFORK is a similar annotation that
replaces a private, anonymous mapping with a fresh mapping of zeroed
memory in a cloned address space. Is it undone through the KEEPONFORK
advice.

This also implements MADV_NORMAL which is a no-op advice value used by
the ltp tests for MADV_WIPEONFORK and fixes a small bug in the order of
error checking on overflow. We're supposed to check for overflow on the
end address of madvise() before checking if the pointer value is out of
range for the address space.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919753
Original-Revision: adb32592fe99b7228ddd036f02660bd3c88137b8
GitOrigin-RevId: fe243e2508a4c400f0e62de20893eca9a66f77a1
Change-Id: I91d36199b9842debd34776bd85a6e20d6eefdf6f