[roll] Roll fuchsia [fidl] Require explicit unknown interaction keywords OOT
This changes the fidlc and replated tools to require all protocols to be
marked open, closed or ajar and requires all methods to be marked strict
or flexible. This was changed in-tree in June but this change will
extend it to petals. Before this can land I'll have to update all FIDL
sources in petals to include the keywords.

This will allow us to change the defaults later, keeping the ABI and API
unchanged.

Original-Bug: 88366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/911314
Original-Revision: 7e3f2c6d245034792e602044d0283ed9df75634e
GitOrigin-RevId: d0d4ffdbed75d4e906d631e8965ed148baf525b7
Change-Id: Ice4a42aaa8f201c1b839ad5af80d09d178d2469b