[roll] Roll fuchsia [bt] Assign dynamic PSMs when requested
Allow FIDL clients of the bredr.Profile API to specify a dynamic PSM in
the ServiceDefinition. When requested, the Fuchsia Bluetooth stack will
allocate the next available dynamic PSM for the service.

Randomly generate the next dynamic PSM and fall back to a linear
solution if unavailable.

Add a testonly hook to the SDP server to return the set of allocated
PSMs. This is used to identify the dynamic PSM.

Original-Bug: 133715
Test: fx test bt-host-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/916172
Original-Revision: 6b7e0ad41d8a9b05d826e1842e6f4fc25e02d818
GitOrigin-RevId: a4adf4c2daedf0d78ae7979125648900d72d579e
Change-Id: Id62975bc3daffc00bc809d4ce67fb49a448c724d