[roll] Roll third_party/pigweed pw_cli: Narrow logic around colors
We would always use colors if PW_USE_COLOR was set to a "true" value, or
if it was not set, or if we're connected to a terminal. Change that
logic to default to using colors if connected to a terminal, and
overriding that default if PW_USE_COLOR is set.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/173232
Presubmit-Verified: CQ Bot Account <pigweed-scoped@luci-project-accounts.iam.gserviceaccount.com>
Original-Revision: 4a65df27bf9e41d5cd92f9dcc5644c4fc6d1dc23
GitOrigin-RevId: fd7c6fee4ebe0077a32a0321c9821280453e9853
Change-Id: I7529cea5f850a6b2f566bdb186d75f3e35a98c27