[roll] Roll fuchsia [audio] Inspect data for dynamic delay
Change 'internal delay' and 'external delay' from single
static properties to initial values, plus current value
and the time of that change. If needed in the future, this
could be changed to an ordered map, with the key being
the monotonic time of change and the value being the pair
of delays (both internal and external).

As a result, Inspect can show whether a driver's delays
have changed.

Test: audio_core_shared_unittests
      Unittests added, plus Inspect shows the changes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921192
Original-Revision: eae5a8b24b00d06408b76f0412b851041d5df3f5
GitOrigin-RevId: 492d94f27f7615024480dd5c2160dca732fb1045
Change-Id: I6ea8502ebe246a5f9447287c5666321958f7f617