[roll] Roll fuchsia [fidl][cpp] add fidl::IsWire<T>
This allows code to specialize based on whether a type is from the
natural or wire bindings. We need to do this because sysmem relies on
internal implementation details of the new C++ bindings to guess this,
and with the removal of the V1 wire protocol those are changing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/923134
Original-Revision: 9e5d02589fa15d456b9a1cb28b1ae061e9efa0da
GitOrigin-RevId: 32cd6744ecf0ce0c2f701116826f9f8a9d9d9612
Change-Id: I6ec19ed855a5ee22c87b3ad3851282bc1e740548