[roll] Roll fuchsia [fx format-code] Fix vendor/ repo formatting bug
I recently introduced a not-yet-functional shac.star file to a vendor/
repository, not intended for it to take effect yet, but it got picked up
by shac's default recursion strategy and broke `fx format-code` for the
vendor/ repo.

Setting --no-recurse isn't ideal because it means that more deeply
nested shac.star files will never be run. There aren't any such files
for now, so it's not the end of the world, but it should be fixed ASAP.

Original-Bug: 82386
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/923592
Original-Revision: cb00f4fa4fbec369de5be4277f1c9bdaf7459ab8
GitOrigin-RevId: 5b22f1c6fe2c32fc2dec24d81e4c497265cdaf1e
Change-Id: If8151f1750905f0fb0dec2b16f121448872e786f