[roll] Roll fuchsia [fshost] enforce tear_down in tests
Add a check to enforce calling tear_down on the test fixture before the
fixture is dropped, and add the call where it was missing.

This also revealed a situation where we were filing a crash report for a
corruption after a normal operation where we shred the keys, so some
refactoring is done so we don't file the crash report. Technically this
also fixes a bug where shredding the keys would break a device, if the
device was configured not to format the data volume on corruption.

Test: fx test fshost_integration_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/922312
Original-Revision: 1e02c9a69695d58bf07254378e6d99efe2d21134
GitOrigin-RevId: d10dae417c1257455ad4b7d6df54baf42345483d
Change-Id: Iecf33b127ee97e3e9f3b8cfe4828bd2cb593dced