[roll] Roll fuchsia [superproject] Roll third_party/pigweed docs: Start the glossary
Creates a new glossary (//docs/glossary.rst) and starts it off with a
definition for "module". Updates other docs to link to the glossary
definition.

Over the long run we may want a more structured approach. Maybe we
store the glossary definitions in a different format (e.g. JSON) to
make it easier to auto-link to the glossary in the rest of the docs.
For now, we should just start with the simplest solution and only
add complexity when necessary.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/172952
Original-Revision: 7ea3c0814e1f56d9b1e3148d490ab1540070f1bb
GitOrigin-RevId: cb4fc7f706f89afd977d4c606ccab336d81ba0eb
Roller-URL: https://ci.chromium.org/b/8768678551862459377
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924172
Original-Revision: 6c91c66ac913cc2737c202be9459a27501da8d35
Change-Id: Ib5e54146571c5b80a20a31891b77cae2ce9fd54d