[roll] Roll fuchsia [bt][avrcp] Set browsed player after browse connection is set up.
After browse connection is set up, the controller sets the browsed
player. If a player is addressed, we set the browsed player to the
addressed player. Otherwise, we set it to a player with the highest
browse capability. The player will be set as browsable if it supports
of browsing.

Browse player is set in the peer not in the controller, so that
different controller instances (i.e. from ControllerService and from
BrowseControllerService) that point to the same peer can have the
already addressed browsed player information.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/830496
Original-Revision: acf79c852ab08bac9cc73f99d23fbb2cd0ecdb50
GitOrigin-RevId: a5a81efbef310ff215cc7ce0da9ff37f5478be79
Change-Id: I18330b1483fdf9ba90f059875e583021521e1510