[roll] Roll cobalt-registry Fix typo in comment
Fixes a typo in the projects.yaml header comment. This CL is also
used to allow https://fxrev.dev/923314 to pass presubmit tests.
Part of the registry parser's presubmit testing includes checking
out the parent registry to test the config change validator.
However since the last registry CL was the CL that updated the
registry to use the new format, the parent registry CL is still
in the legacy format, so the registry parser in
https://fxrev.dev/923314 fails to parse the parent registry.
Adding an essentially no-op commit to the registry will make
both the HEAD and parent commits both be in the new registry
format, so that https://fxrev.dev/923314 will be able to pass
the CQ presubmit tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt-registry/+/923899
Privacy-Approval: Cameron Dale <camrdale@google.com>
Original-Revision: 25b845c710ae02a97619dc61899ff3c5a6877d5a
GitOrigin-RevId: a107b29c06fa2c896450dfe6d21bcabd4100ceda
Change-Id: Icbe2d3c8d7db0ed35ee43b193b4f20f22d2d312b