[roll] Roll fuchsia [cpu-trace] Migrate To a Regular Component
cpu-trace doesn't need to be a driver anymore. It originally had to be in
order to get the root resource, but that is no longer the case.

This patch modifies the architecture of cpu-trace to be a
fuchsia.perfmon.cpu fidl server. Provided a DebugResource, it provides
an interface to the mtrace syscall. Then we make it a child component of
the frontends: cpuperf_provider and the cpuperf shell package.

Original-Bug: b/294603854

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/901300
Original-Revision: 15f5cf62d9a91703925d8fa321df85a5c44645b1
GitOrigin-RevId: aceefe7267555e723c6ac01da2cfa571af41f180
Change-Id: Iefc0529e79f4ed2667c189c1fc5955b440a861c9