[roll] Roll fuchsia [archivist] Fix bug that was returning escaped monikers
Colons in monikers shouldn't be escaped. This is only done in component
selectors to remove ambiguity with selector segment separators.

Fixing this led to the need to update some goldens in particular in
iquery as well as ensuring that the escaping remained present when
we were dealing with selectors.

Run-All-Tests: true
Original-Fixed: b/302150818
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/920352
Original-Revision: 80344a494e90b710a8c6f65fa1e249e4ec8004f0
GitOrigin-RevId: 35205576f315fd510289f7548e9f10db84295dde
Change-Id: I110c84e9cac263f013a92be8d129e19d8c8a8f4d