[roll] Roll llvm-project/libc [libc] Correct 'memrchr' definition and re-enable on GPU (#67850)
Summary:
This was disabled on the GPU because it conflicted with the definition
in `glibc`. According to information online and in the `glibc`
implementation, the first argument should be a `const void *`. Fixing
this resolves the problem when exporting this to offloading languages.
GitOrigin-RevId: c90e6bf567068c3e8d046ae4862585647bcbc570
Original-Revision: 4220b39ff9df454d109f943182998959c5beacc7
Change-Id: Ib65fbfc00138b62c14ea699ab19ad8b2b70653d3