[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Correct 'memrchr' definition and re-enable on GPU (#67850)
Summary:
This was disabled on the GPU because it conflicted with the definition
in `glibc`. According to information online and in the `glibc`
implementation, the first argument should be a `const void *`. Fixing
this resolves the problem when exporting this to offloading languages.
GitOrigin-RevId: 12587b2be0b617284116209fc7927c3dababdd51
Original-Revision: 4220b39ff9df454d109f943182998959c5beacc7
Roller-URL: https://ci.chromium.org/b/8768566334623696465
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924715
Original-Revision: 37892e0a0eff9388d46a825bfdce764283201d2d
Change-Id: Ibdad8474c4e98203fcb21d27c3b8ef1c909b3cdd