[roll] Roll fuchsia [fidl][compiler] Rework compiler diagnostics tests
This introduces ExpectFail and ExpectWarn methods on the TestLibrary
class and an ASSERT_COMPILER_DIAGNOSTICS helper macro that compiles the
library, checks that the diagnostics match the expectations and prints
any differences.

With these a test can assert more precisely which errors and warnings
should be generated by a compilation. Without these, when a test that
checks compiler errors or warnings fails it can be hard to understand
why.

This updates a couple of tests I had to address for unknown interactions
to ASSERT_COMPILER_DIAGNOSTICS but leaves most untouched. As we
experience test failures related to diagnostics we should update those
tests to use this approach.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924088
Original-Revision: 4ccef4be5740b601a7e0c758840785d48e46da4f
GitOrigin-RevId: 4a4f116f55166fff7b60b86a2b933bd8c98df051
Change-Id: I8b709157474848f26c8b1fa7c767c5fd29c670bb