[roll] Roll fuchsia [amlogic-display][vout] Remove "VoutType::kUnknown".
kUnknown used to indicate that the Vout is not yet initialized
with any type of output. Now that the output type is always set
up at Vout class construction, it's impossible for a Vout to be
of "kUnknown" type, so we can safely delete it.

Test: no-op. Dead code deleted.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924250
Original-Revision: 6e02244dc238eb3c910947c51a6244181ecac315
GitOrigin-RevId: 837d35289104ceb096dd1a8d7de8cc1001d09059
Change-Id: I1d5924f90bb82a219091f986bf60968b89191c66