[roll] Roll fuchsia [audio] Inspect for min lead time changes
On a per-renderer and per-capturer basis, change 'minimum
lead time' (or fence time for capturers) from a single
static value to a trio of values: initial value, current
value, and time that value changed to current value. If
needed, this could be changed to an ordered map in the
future (if the full history is needed).

This should aid in the diagnosis of render/capture issues
where we observe the client's packet handling getting out-
of-sync with audio_core's expectations (as shown by
renderer underflows).

Test: new unit tests, plus Inspect shows the changes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921193
Original-Revision: 42d39da18fb6bd6bea839caef48bacf3f8b56f4b
GitOrigin-RevId: d259af1641f39a118feb42246bdf9d645615bfa7
Change-Id: Ib35d0117577d5f60920ed9734956594b3349c5a7