[roll] Roll fuchsia [netstack3] Allow adopting more state in Locked
Working with Locked when we have disjoint state did not benefit from the
lock level L being upgraded together when acquiring locks for each piece
of state.

This is mostly useful when we have resource state behind an Arc (like
devices and sockets) and their internal locks are part of the global
lock ordering graph.

While I'm here, make all the parts of Locked that are opinionated about
the storage agnostic which allows for
a) now, having Locked own a tuple directly without requiring a temporary
variable and a way to escape the internal ref;
b) future, implementing traits on Locked<X, L> where X:
Deref<Target=StackState> which _may_ provide a path for ownership-based
APIs, if we so desire to move that way.

Note that UnlockedAccess is still opinionated about acting on references
because it does not want to inherit the lifetime of the mutable
references like all the other functions do.

The xxx_with family of methods also allow for a more ergonomic future
where the lock levels are annotated closer to the data they protect as
opposed to globally on SyncCtx.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/925132
Original-Revision: 7821319fe4ac360859026edd959b9a96e9abe5b3
GitOrigin-RevId: dfbb2987ed7ef6ccac9cea7752fbf17a7d34389a
Change-Id: I6479fc02421f9386d0c3ad2c1b9e757b5796d4a4