[roll] Roll fuchsia [sandbox] Remove flags and target_moniker from Sender/Receiver
Sender and Receiver now transfer only Zircon handles, not Messages.

Flags are not needed because Sender/Receiver are only used to route
protocols, and flags are not used when connecting to a protocol.
Target moniker was used in a previous version of routing code to launch
the target component on connect, but is now unused.

Removes a dep on the moniker lib.

Original-Bug: b/298112397
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/925295
Original-Revision: 35f513b3d7e60425b6feb76eedea2c12eeb9bfe9
GitOrigin-RevId: e39850c2dde708ebbba3a28e9e7c9cbea3e4b3f8
Change-Id: I06596be95638bed8a70636a6d1a3e05f15e6abaa