[roll] Roll fuchsia [docs] Remove references to RFC-0083 and add note
This CL changes docs to reference the main "FIDL versioning" page rather
than "RFC-0083: FIDL versioning". It also adds a note to the top of the
RFC linking to the more useful documentation pages.

RFCs are historical documents and not meant to be kept updated, but
people end up on this RFC (hopefully less often with links to it fixed),
so it's useful to point them in the right direction.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921134
Original-Revision: 818df65453cad51dd8351cd88074d606a00575e8
GitOrigin-RevId: a9070f7a4587025025a205888c7c5c3f56ed1cb8
Change-Id: I36c8eafe2a7aa6a53521a7c474f7dc42b2b49078