[roll] Roll fuchsia [ffx] unbuffer reading of ssh stderr
Diagnosis of some bugs requires accurate timing information from the
"ssh -v" output. When the output is buffered, the timing information
is lost.  This change removes the BufReader from the stderr line-
reader, replacing it with one that reads bytes directly. (It may be a
bit slower, but its use is limited to stderr, as well as getting the
first line of stdout. It should not affect the steady-state of a working
connection.)

Original-Bug: 301312339
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924399
Original-Revision: 4a29b2b88d173c8b6d32274a136847a1a1b1cc0c
GitOrigin-RevId: cc535c83bdf3ea86211905cd6107b7e0c72786e3
Change-Id: I97ae4d21d86d8901b09859409110bf036b3c111d