[roll] Roll fuchsia [starnix] Fix remote binder interruption mechanism
Remote binder needs to be able to interrupt a process waiting on a
binder call when a task needs to be freed to execute a taskless
operation (like setting the VMO, or creating a new thread). There was a
race in the current mechanism which could create deadlocks.

This update binder to let client interrupt the wait reliably.

Original-Bug: b/296893420

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/925112
Original-Revision: 51666b388186eee7aeb2519b5ac8fbccbe4e5c39
GitOrigin-RevId: 5ee54ff96972ec0f983faf16a8e787d165c2900b
Change-Id: Iac7ff0acb00545f42e3be53305f0cba3a423c277