[roll] Roll fuchsia [shac] Move clippy logic into shac
The staticlints tool is deprecated in favor of shac, so move over the
clippy logic. Now it will run as part of the `static-checks` builder
instead of Tricium, but will still emit Gerrit comments.

The clippy file-parsing logic is complicated enough that I left it as a
separate Go tool instead of moving it into shac Starlark code. However,
this change removes the restriction that the implementation must be
entirely in Go, so in the future we could delete the clippy-related Go
code entirely and instead have shac call into the same Python logic as
`fx clippy`. That's a bigger and riskier change which is why I didn't do
it here.

Original-Bug: 82386
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924613
Original-Revision: 4b7e6b12c777a8df0adf9588f2886e98330e446e
GitOrigin-RevId: 64c39f2dad554d111e6e99a4016720ee8cafcc50
Change-Id: Ia64d91550b2232c3828365535624d95731ea2e1f