[roll] Roll fuchsia [pkg] Create PackageResolver-ota capability
This fuchsia.pkg.PackageResolver capability, exposed by the
pkg-resolver, is the same as the original, non -ota,
implementation, except that the resolved packages are not added to the
dynamic index. This capability is only to be used by the
system-updater, which will protect the packages from GC by adding them
to the retained index first.

Later CLs will:
  1. Migrate the system-updater to this new capability.
  2. Disentangle the retained and dynamic indices, by having
     membership in the retained index no longer prevent addition to
     the dynamic index.

At which point the only client of the dynamic index will be non-OTA
package resolves and we can replace it with open package tracking.

Original-Fixed: b/293331949
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921312
Original-Revision: 8c15aaaa40f149c7651b2c6e20537b6a76ad1796
GitOrigin-RevId: ebea5165eb9b4cfcdbf340bd016f74f6b5061aae
Change-Id: I485603c046fc019f9d41dedf4444f6cbef635edf