[roll] Roll fuchsia [netstack3] Add instantiate_ip_impl_block macro
Introduce macro that repeats an impl block for IPv4 and IPv6. This
gently sidesteps the lack of specialization features while still
allowing effectively complete implementations for traits that are
generic on Ip.

This provides a meaningful and useful workaround to reduce code
duplication in Netstack3 due to the virality of non-Ip-complete
implementations as we walk up the dependency chain.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/925472
Original-Revision: cd6bba3d9fac13abd97312320af441801c1555ba
GitOrigin-RevId: b8bfe4c276e4971be72dbc707932369db5d0aa7e
Change-Id: I32eca7a1cf892e7342a2fbac7c53f0396cef7617