[roll] Roll fuchsia [netstack3] Fix error code for changing multicast membership
Before this change, we return a single error code if the membership
is not changed. However, we should differentiate between joining an
existing group and leaving an unjoined group and return different
errors. This problem was not exposed previously because the NTR test
mistakenly launched NS2 even if the intention was to launch NS3.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/926016
Original-Revision: 5d28dad244ad6f46a9039e5cb4b8b648877f78ed
GitOrigin-RevId: e797739b79d078ee6ad777eea9b163de64354af4
Change-Id: Ib9df230b7a26561fc1acccd6c84ae41c06b274bf