[roll] Roll fuchsia [kernel][tlb] Fix bug in shared entry invalidation
Fixes a bug in which TLB entries that are part of a shared address
region are not invalidated correctly. The code originally attempted to
use an `invlpg` instruction to invalidate entries of a given address
across all PCIDs. However, this is not actually what the `invlpg`
instruction does; from the intel documentation:

"The instruction invalidates TLB entries associated with the current
PCID and may or may not do so for TLB entries associated with other
PCIDs."

This was leading to certain tests accessing stale TLB entries and
crashing. This CL fixes the issue by upgrading all TLB invalidations of
shared entries to a full TLB shootdown (though it is important to
highlight the fact that this may not always shoot down all global
entries).

Original-Bug: 132980
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/926019
Original-Revision: 36cfee916f6c5479fa8fe093c68b156bfa480a78
GitOrigin-RevId: cd4674c22dbf2a99c60c2eb6c2bc35c91dcf5a14
Change-Id: Ic143c181beb504da1a8863854ad2d58e49ff6bb2