[roll] Roll fuchsia [ffx][target-collection] Transient and enabled target flags
The enabled flag indicates a target was previously used.
The transient flag make a target transition to disabled when it expires.

None of the restrictions associated with the enabled bit are present in this CL, just the state tracking.

Original-Bug: b/299129117
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918955
Original-Revision: 62365cd5fddc829c294a775ef7e0a8ea82f050b2
GitOrigin-RevId: 0e8abd2bb51a253601c73c30fe5656884aefcb59
Change-Id: Id79893c48107b85abf009b8860be1c0957716922