[roll] Roll fuchsia [scurtiny-x] Drop internal model of component manifests
`scrutiny_x::Component` behaves as a component manifest that stores the
identity of the package the component happened to be loaded from. In
other areas of the API, we have recently decided that the object from
which another object was constructed is information easily retained by
clients that need to maintain such a relationship, and need not be
exposed via a Scrutiny X API. This change aligns management of component
manifests with that philosophy.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/925093
Original-Revision: 87f9fca5feabc4d35309ce5931bfd198c799a55d
GitOrigin-RevId: ef3e3cc4c4154ce4cf641e50fcdbba3ebb1d6f28
Change-Id: I1cfbc2dbdaaa1225d0fad14c4846f825fdc7bc2f