[roll] Roll fuchsia Reland "[kernel][arch] Introduce memory types to paging traits"
This is a reland of commit 247c60cab19a9efedbd2af7c6f11d8679136c232

What changed: PagingSettings was made to be a definition of the traits
themselves, allowing control over the default memory value. The previous
approach of using a std::optional<MemoryType> to allow for a safe
default-constructed value ironically wasn't used safely by the author
and led to unintentional dereferencing of the null state (as well as UX
complications).

Also, the paging trait definitions are moved to the top of each
<lib/arch/$arch/page-table.h>, prioritizing the definition of the main
attraction in that file and reducing the expedient definitions made in
the library scope.

Original change's description:
> [kernel][arch] Introduce memory types to paging traits
>
> This change introduces memory type support to the paging traits. This
> provides the functionality to configure pages as normal memory or MMIO.
>
> Original-Bug: 129344
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878980

Original-Bug: 129344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/923652
Original-Revision: 0bdab47fcd5fbe674d8a13d23a4f206d3d9d5591
GitOrigin-RevId: b2d48a33a3343f7b5c14280eb6101b6e282b5bbf
Change-Id: I3bb771f9b27dd3d84e81b71cfc53651955955259