[roll] Roll fuchsia [ffx][goldens] Don't check when building separated tools
This makes it so that the cli golden tests aren't done when the
`ffx_build_dual_mode_plugins_as_subtools flag is set, since the SDK
built when this flag is incomplete (does not contain the tools that are
built separately).

Also bumps the number of "internal" tools expected in the help-json test
down to 5 because once we migrate things out of the sdk that's about how
many there will be, and for now that helps the test pass with the flag
on.

Original-Fixed: b/303319158
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/926736
Original-Revision: cd33eb9d4f53923fa4d1e3a309e08cf7440f334c
GitOrigin-RevId: dff1db58e27af89d3c7ce7fe0e68fd0d009b9ac7
Change-Id: If5189b99021df4d27525e6cfa3c4e8be4a9692b3