[roll] Roll fuchsia [driver_manager] Fix flake in restart_on_crash
Fixes a flake in the restart on crash test that occured when the node
closure handler was called later than the FinishRestart function. This
was causing an extra removal to happen. The fix is to reset the
node_controller manually before we try to start back up.

Original-Bug: 134678
Tests: Added coverage for node/binding races in unit test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927221
Original-Revision: cc511b94219745c5eb80b1eed210501a4d4e4911
GitOrigin-RevId: 9edcd3e08d03b8ece494b388a6a7d0ecc8a19643
Change-Id: I181319b8dd3e620890f5e91b63a2ce8173f98180