[roll] Roll fuchsia [ctf] Remove the allowlist
This CL removes the allowlist and the GN templates and scripts
that implement it. The allowlist is being removed because it
does not prevent tests from depending on things that aren't in
the partner-sdk category and adds toil to test authorship. We
will need to replace it with a different mechanism sometime
later. Additionally, this allowlist was a large part of the
reason the ctf_* templates exist. Most of them now only forward
arguments to their non-ctf counterparts and will be deleted in
a later change.

What is removed:

* ctf_element.gni
* ctf_version.gni
* write_ctf_file.gni
* verify_deps_in_sdk.{gni,py}

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/925192
Original-Revision: fbd02c5ebbf5652d4892d162ffbb5f01fd51443a
GitOrigin-RevId: 344850edb13ba1dc1ef98594e07074e06fbb95ba
Change-Id: I81ad724052c98a333201d738ec3a9ec4b16432c6