[roll] Roll infra/recipes [rust] Disable ThinLTO on beta builds
Use of LTO has led to miscompiles of our compiler in the past, so it's
not a great idea to use it for bootstrapping.

This is a reland of commit e0e48afb8f8339e75080f69ec8417077948d34cb.
That commit broke the toolchain builders because disabling LTO exposed a
pre-existing issue in our build that LTO had been hiding. The issue was
that we were linking libc++ into the compiler driver binary but not
libunwind, which libc++ depends on.

In the cases where we link libc++ we need to let the driver (clang++)
use its knowledge of other dependencies like libunwind. This is
accomplished with the rustc flag -Cdefault-linker-libraries=yes. This
prevents rustc from passing the clang flag -nodefaultlibs and trying to
replicate the logic inside of the clang driver.

Linking to libc++ only occurs when building for host (needed by
libLLVM).  Generally the flag shouldn't be harmful, so we put it in
RUSTFLAGS for all non-Fuchsia targets. It wouldn't affect Fuchsia
targets anyway since rustc invokes lld directly for those, and it
doesn't have any default libraries.

At first I was worried that the target-specific RUSTFLAGS would not
apply without -Ztarget-applies-to-host, but it did. That problem only
occurs with outputs that cargo considers "host" like proc macros and
build scripts, but not "normal" outputs that happen to target the host
platform.

Original-Fixed: b/302039535
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/923097
Original-Revision: 611455e06286befcc085db62be1d42384f03cd2b
GitOrigin-RevId: 8a804424a91348777e309150cd330e42f0b123af
Change-Id: I3c788e98494d90ff7901ddcdeb88ff184496489e