[roll] Roll fuchsia [starnix] Don't attempt to sync notifier and notifiee
...of a `PortWaiter`.

In commit 3027cfc2faee116546f84aa918951299ac3f860f (fuchsia CL
https://fuchsia-review.googlesource.com/c/fuchsia/+/924473), it was
discussed whether the `PortWaiter` should specify a happens-before
relationship between a notifier and a notifee. To accomplish this, the
`has_pending_user_packet.swap` operation used Acquire/Release ordering.
However, even with such ordering, the `PortWaiter` cannot guarantee the
happens-before relationship in all cases.

This is because immediately after an atomic load with Acquire ordering
but before the waiter processes any triggered events, a new event can be
triggered and there will be no following atomic load to synchronize any
atomic store (with Release ordering) with. Because of this, the only
way to synchronize a waiter with a notifier is to have the waiter
perform the necessary synchronization itself.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927213
Original-Revision: dd27baa1388dfbb072ceeeb6ace2f81b36277905
GitOrigin-RevId: 2bbf4ea4f1401114adac65c263fe5302d2bc1f9c
Change-Id: I97726f00bb6a23a8e130f08833b7fb315ecfde9b