[roll] Roll fuchsia [testsharder] Run all tests if no affected tests provided.
If there was no affected tests file provided, that more likely means we
were unable to determine affected tests as opposed to there being no
affected tests, so we should run all tests. This is generally the case
but was missed in the case where affected-only was true (which isn't yet
used anywhere).

This also fixes a bug where an empty affected-tests file was causing all
tests to be marked as affected.

Original-Bug: 133838
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/923296
Original-Revision: 2d2122dd3054e1087ec89a828557d48574bda91d
GitOrigin-RevId: 5533900be6813bc85313eafe35af7b29de2d487b
Change-Id: I3c7fc745303bacccc8112f891d731d56ab0fadb8