[roll] Roll fuchsia [storage][benchmarks] Add ReadSparseCold benchmark.
The original ReadRandomCold would make a file and read back the offsets
in a shuffled order. While this might seem like it tests page fault
speed, due to the whole file being consumed throughout the test the
readahead benefits were still significant and favour the filesystems
with larger readahead. In the case of small random page faults readahead
should be a hindrance instead of a benefit.

The simplest answer is to not read back the whole file, but in order to
take many samples the file needs to span a much larger space. Too much
space not only makes the test setup take a long time but increase the
space required on devices running the benchmarks. In order to get the
same amount of samples while bypassing this readahead benefit we'd need
at least 256MiB file versus the current 8MiB.

So we use a sparse file, which allows us to only use 8MiB of space, and
some order of magnitude smaller extra overhead to store the many extents
that the blocks will now be indexed with. This additional metadata will
also create extra steps for the filesystems since they will now need to
look up different extents for each read, whereas previously filesystems
were likely just creating single giant extents for the whole file.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/922952
Original-Revision: 75ce198cca4d10926b63107d765a080cd054c9f1
GitOrigin-RevId: 1c9789087d27466ffb0369336f427a637452eb82
Change-Id: If9232c61d94dd191c5a69db2a495e01fcbeb721f