[roll] Roll fuchsia [amlogic-display][vout] Ignore `mode` if EDID is not present.
Per definition of `edid_present` in `AddedDisplayArgs`, for a given
display, if its EDID information is not present, the meaning of
the `mode` field of DisplayConfig struct will be undefined and the
driver should ignore it in CheckConfiguration() and
ApplyConfiguration().

This change adds a field `ignore_display_mode_` to AmlogicDisplay
so that it can ignore the `mode` field when checking or applying
configurations.

Test: no-op expected.
Test: display-tool squares 1 on Nelson (MIPI-DSI)
Test: display-tool squares 1 on VIM3 (w/ MIMO HDMI display)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924397
Original-Revision: 67427c3e36352ea8e169b7eec74b50dc56fbab93
GitOrigin-RevId: 03182b6b407d98efe8ffce98a389df37bf2c399c
Change-Id: Ifeae5463de062d18b76bfd0146d696fe7e073266