[roll] Roll fuchsia [overnet] Delete examples, hoist, and host tools
All unused stuff, except perhaps for the examples, which used to be
exercised as part of our testing strategy. They work the handle types
specifically, but our fidl tests do that, and the connections end to
end, but ffx owns the production version of that and exercises a more
worthwhile version of it.

Original-Bug: b/287694609
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927133
Original-Revision: 05ab5e4a5e4412d308f9621188aff4b8714f22e5
GitOrigin-RevId: 8dbc0b2f9c4a310e22ab466082eb73430e166d38
Change-Id: I09dfb9cd291b199564d897ae91497196fa8ead15