[roll] Roll fuchsia [ffx][compat] Fixing golden file generation
This is fixing the golden file generation when a new
sub command is added.

The problem is that the loop creating the golden comparison targets was missing dependencies on the sdk - where ffx comes from so it would use the contents of the SDK before it was finished building the sdk, picking up the old command info.
so now the sdk has to be built, then the top level command list, then iterate over that list, vs. using a stale list.

Original-Bug: b/303542110

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927835
Original-Revision: 4f287b9ac83c29dea500ba372b4cbb623a5eeb01
GitOrigin-RevId: be2f5093ebbd74da4568a3a67bd382f2480f1504
Change-Id: I936221b85d3f9f10c52c56605c309a6228448a75