[roll] Roll fuchsia [amlogic-display] Skip duplicated display modeset in AmlogicDisplay.
AmlogicDisplay handles the upper level logic of
fucshia.hardware.display.controller.Controller interface.

Skipping of duplicated modeset requests is part of the upper level
logic. This change moves it from Vout to AmlogicDisplay, so that
AmlogicDisplay compares the new display mode with the mode applied
previously, and calls Vout::ApplyConfiguration() when it needs to
apply the new mode. Vout::ApplyConfiguration() now unconditionally
apply the new display mode to the display hardware.

Note that Vout still needs to store the currently applied
display mode so that it can be used to determine the display size.

Test: "display-tool squares 1" works on VIM3 (w/ MIMO HDMI display) and
                                        Nelson (w/ MIPI-DSI display)
Test: "display-test --mode-set 1 3 --num-frames 1000"
      (changes display mode) works on VIM3 (w/ MIMO HDMI display)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/915698
Original-Revision: d7d48325a498c1923346cfa10d04157783178851
GitOrigin-RevId: f0c573a53b45e82fdc25931096d8692c16a9d31b
Change-Id: I3e8e6de907e92a8a90f74ffacd91cf05d221fb92