[roll] Roll fuchsia [bazel-sdk] Add a repository default api level flag
Adds the ability to set a flag in a .bazelrc file which will serve as
a fallback api level if one is not set on a package or on the command
line.

This CL adds 1 new flag and updates a previous flag. Each of these flags
will be put behind a flag alias to make them easier to use.

--@fuchsia_sdk//fuchsia:override_fuchsia_api_level allows a user to
override the api level for a single command.
--@fuchsia_sdk//fuchsia:repository_default_fuchsia_api_level allows for
a fallback api level to set on a global level.

BUG: b/303684806
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927841
Original-Revision: 7c68f71e03c0dbad93fc15d9b5293ce84744d533
GitOrigin-RevId: dbbb9165bd5bf12d1ff8c90db514d41e489f15b3
Change-Id: Ic6c3170508f1709b77eed5871b3f467b05b8d60f