[roll] Roll fuchsia [bringup][assembly] Remove special-casing to add base drivers to bootfs
This removes the special logic that copies the contents of all base
drivers into bootfs on a bringup build.

The device-enumeration test is changed to include the packaged drivers
when compiling the test package, but not when compiling the bootfs
test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927844
Original-Revision: 052663cd86b139aa93eaf0b819c0b8e7663b08f2
GitOrigin-RevId: f6a092f941be29e64fa7b9d17e728a16d3c220a5
Change-Id: I93d88aa4607f8dda773249139db5e4f65b795fb7