[roll] Roll fuchsia [starnix] Fix bug in read_c_string_to_vec
This CL fixes a bug in starnix that would cause long c strings to be
corrupted.

When reading from userspace, if the user string was too long, the code
would resize an internal buffer and continue filing that buffer from where
it left off, but it would not advance the string pointer. This would
lead to a string prefix being repeated multiple times over and over.

For example, if a string was exactly 0x1000 bytes long (not counting the
NUL-character), the code would read the first 0x1000, note that there is
no NUL char and resize the internal buffer. The next read would read
data into the new part of the buffer, but the source will still be the
origin of the string.

The CL also adds a test for this scenario: it will write a random c
string of 4 pages length and then read it back, expecting it to be the
same.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/928492
Original-Revision: e6258804c2990a494dcda2bcb0148e96d36f08bb
GitOrigin-RevId: 1799372d183ae60828f308aca493d3ce0ae47fed
Change-Id: Id4f6304ef70a5f46aa8de29070be98bd23a1acd6