[roll] Roll fuchsia [ffx] Reap ssh processes on failure to connect
A recent change caused the code that reaps ssh processes to be
skipped when there was an error connecting to the network. The
result caused unlimited numbers of ssh connections to be created,
then left in a defunct state, eventually leading to a system failure
as the process table filled.

We now kill the process when the connection cannot be made.

Fixes: b/303721424
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/928672
Original-Revision: 6132bd138d24d09d75066263623584539213bd09
GitOrigin-RevId: f99a589ed7e2b42b1db9535711b0a9619fcd72eb
Change-Id: Ie1bd5acf82c7a46106c99107029d0714d27562d7