[roll] Roll fuchsia [ffx] add net diagnostics request to e2e_emu tests
A flake that seems to be related to netstack can't be diagnosed
with normal, networking-based tools. Instead we'll add a temporary
hack for a set of tests that seem to exercise the failure,
which invoke some target-side diagnostics that can be inspected
by the netstack team.

BYPASS_INCLUSIVE_LANGUAGE_REASON=TERM:dumb is required
Original-Bug: fxb/133234

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/928334
Original-Revision: d2d86b36fbbea340f0918a9abf9cf532b5015124
GitOrigin-RevId: 82b78f89d4878789c0a3ca8360d442fbc7179834
Change-Id: I7b7461cf7b7be1113fd7c09f65e7302beea6c1ee