[roll] Roll fuchsia [x86][phys][boot-shim] Fix build of tests on arm64 hosts
This test depends on a UART that is only defined on x86/x64 or if
UART_ALL_DRIVERS is defined.  The latter appears to be an affordance for
testing, so let's enable it on this test which is now being built on an
arm64 host for the first time.

Original-Bug: b/283475308
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/928468
Original-Revision: 18940849686d9b5491d1a1219c1bd09152151e32
GitOrigin-RevId: 1aafa151a701669e339c4bece47c5ae0468494f8
Change-Id: I35f329a858a18c5e5531e19ef9ee89eb6ef9d415