[roll] Roll fuchsia [starnix][lock-ordering] Introduce lock-sequence crate
This is a modified fork of netstack's lock ordering mechanism that
allows to statically enforce lock ordering by adding a new context to
the locks, indicating the "level" of the lock. The order of the locks is
then expressed as relation between these levels.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/893918
Original-Revision: 5da4a0e615e03ccee0e2bd291b3f395dca52cea9
GitOrigin-RevId: c155c4625152d461e95c0ad609018c2d4bcb7dbf
Change-Id: Ia68545e963063a62dc5b2381f56cfb3fa2de1019