[roll] Roll fuchsia [wlan] Set @available for fidls in partner sdk
The FIDL protocols and definitions exposed via the SDK need to carry an
API level for when it was introduced. Ones marked as HEAD will not be
made available in the SDK.

This was caught when the BazelSDK moved from using HEAD to API level 15
for OOT driver development.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/928576
Original-Revision: 44456e32ed6d0ee97d45cd1459aa032ac9da8cbc
GitOrigin-RevId: 3af8c9c1a6e802e4c9c84b9cc6ec4199cc88fa3c
Change-Id: Icb8f2c1b03dbffbdd272bfe1b1f25b6d390585a3