[roll] Roll fuchsia [bazel-sdk] add component info to fuchsia_package provider
The FuchsiaPackageInfo did not carry the FuchsiaComponentInfo along which
means that our workflows need to depend directly on a component if it
needs to get any specific information about the component like if it is
a driver or not. This causes problems because it forces the component to
be analyzed in a different configuration from the package because we
handle transitions at the package level. This forced us to put
a transition on the workflows which can cause us to run analysis even
more than we should. This does not fix the transition problem but it
sets it up to being fixed later.

BUG: b/303681889
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929115
Original-Revision: e1dd8a962d15143f233c93d26167650166c0ad12
GitOrigin-RevId: 5a484938e46fd36ca4c8347e37b46d3a074fd877
Change-Id: Idc2220e8dbc474a008672e1886d0c6e08a7995ed