[roll] Roll fuchsia [ffx][resolvers]: Make Resolver not require flash manifest
Resolver now takes a root_path and does not implement the
ManifestResolver trait.

Added new struct FlashManifestResolver in the manifest mod which embeds
the Resolver and implements the FileResolver and ManifestResolver traits.

It's `new` function asserts the path passed to it is a file.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927833
Original-Revision: 72c947c2b45363990bd3d0f86fe60e087bccd551
GitOrigin-RevId: e5b7362c22c9ea8349edfe7f6da429ed929972c6
Change-Id: Ieb98fc87c44a9ef5b670aef49a7c5b2608f92872