[roll] Roll fuchsia [wlan] Prefer MacAddr and Bssid types over [u8; 6] in WLAN code
WLAN recently began depending on redaction of unhashed MACs logged in
the System Log. This change aims to accomplish some overdue cleanup
around general printing of MACs in the System Log.

Primarily, this change aims to significantly reduce the usage of
[u8; 6] in WLAN code which accomplishes the following goals:

- Reduced chance of debug logging a [u8; 6] containing a MAC address.
- Reduced chance of using a MAC instead of a BSSID. (MACs and BSSIDs
are morally the same, but the BSSID is typically special because its
tied to the AP a client connects to.)
- Discouraged creation of new helpers to print MAC-like data.

Original-Bug: b/297365339
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924793
Original-Revision: 6a19ea6ea1c78b04ce39fb42641920159281547a
GitOrigin-RevId: d15ec5a1983c35ac22cf9a6a07659c31ec64cc51
Change-Id: Iab72e6c4bd32c367f677665a7208ae48de745a63