[roll] Roll fuchsia [fxfs] Precommit pages for blob decompression buffer.
We do a lot of inline page faulting while moving data into the buffer.
We should fault them ahead in parallel with the read here to make it
essentially free. This is handy for things like meta.far files that will
decompress a huge amount of zeroes which will write out really quickly
and the time will be dominated by page faults.

This is harder to do in other places and actually harmful, as buffers
that are reading will be committed elsewhere in the stack anyways, but
you also can't easily call read into a buffer while you commit on it due
to borrowing issues.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/926997
Original-Revision: af8d6216f424d26ddb9674ef54f4b40082ca25f0
GitOrigin-RevId: c2d5d2aa45ff23989fa31e210d9efb3f60f0cb1e
Change-Id: Ib0595f648802543d3bb889c6add321ccdc87d370