[roll] Roll fuchsia [rcs] Support connecting to incoming namespace of core/toolbox
This change fixes a bug in remote-control which prevented the ability
of connecting to protocols in the incoming namespace of a component due
to protocols being nested under a svc directory inside the incoming
namespace.

When using OpenDirType::Namespace, the namespace directory looks as:

```
ns/
  svc/
     fuchsia.my.Protocol
  pkg/
```

So requests come for `svc/fuchsia.my.Protocol`.

Additionally this change introduces an API in the rcs library to connect
to protocols in the incoming namespace of core/toolbox.

Original-Fixed: b/304302408
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929192
Original-Revision: 58a192552b149fe4d0b0de8649b6e56dfe6028cd
GitOrigin-RevId: a95c1b4aefd6a954e77ab0f1603ffcb309ba4ee0
Change-Id: I65a9a3d6391636fd6273c70223b82d7bdd78bb24