[roll] Roll fuchsia [vim3] incorporate common/ar64-common and flatten remaining items
This flattens the contents of //boards/common/arm64-common.gni into
vim3.gni, and removes any +/- logic that was a result of the basing
of vim3 on arm64-common.

The remaining items are re-arranged to make clear which are left to
move into assembly, and which are redundant with inputs that are in
assembly, but which are still used by targets in the build.

Original-Bug: b/302567529
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/928464
Original-Revision: c8e48c61a54b6e7e36ca0b5d162483b42b325074
GitOrigin-RevId: 76302c4a638fdec17d17ff7cade8ac3b6016c02d
Change-Id: I6270ad2f2b34bfc94bb7c40c34e0f7b20f4d19d8