[roll] Roll fuchsia [starnix][fuse] Workaround b/297439724
Instead of trying to block the system task when closing or flushing a
fuse file, we create a new kernel task and block that task instead. This
works around the `assert` that is triggered when multiple threads try to
write to the run state for the system task in an inconsistent way.

Original-Bug: b/297439724
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927174
Original-Revision: 78d56d84df0e129b84fc54327a95bf6201b81d57
GitOrigin-RevId: 7231250334b55cd43e8843f95d1f9b9fe1b63d63
Change-Id: Ie5fff49d5279e492b467b93c155367ef6565c794