[roll] Roll fuchsia [devicetree] Translate address based on ranges
ranges property in the parent node contains address mapping. This needs
to be taken into account when calculating the mmio address.

This CL also fixes reg property parsing for nodes in which reg is not an
mmio. This is currently decided based on the size-cells.

Original-Fixed: 133156
Test: mmio-visitor-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924020
Original-Revision: 4efd07013650dcbff63b7ff5dfc59ecc239c6b62
GitOrigin-RevId: e5498240ba0932c0ec64b8102ee91d716783b569
Change-Id: I20d20252231253076b8053bb70b78353733277b6