[roll] Roll fuchsia [docs] Fixing reference links
This fixes markdown reference links (links with [] that are defined
later in the file).

This is in preparation to turn on doc-checker checking of reference
links, so it will be harder to write updates that have broken links.

Original-Bug: b/289246762
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930273
Original-Revision: 12ded5c4ea047c808eb5f8fd1fe1598ea808c521
GitOrigin-RevId: 8817a1302c7b9d75dd64e1639a59022ec0a0765b
Change-Id: Iaee59b7705bf53129670214333d754426723b91f