[roll] Roll fuchsia [bt][avrcp] CT uses GetItemAttributes when browsed player is set.
According to the spec, controllers should use `GetItemAttributes` browse
command instead of `GetElementAttributes` control command if browsing is
supported on target side. Change the controller service's
`GetMediaAttributes` method, so that the above is implemented. As said
in the spec, the CT registers to uses the UID of the currently playing
track and use in the scope of the Now Playing folder.

Original-Bug: 105465
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845116
Original-Revision: dec9fc06cde25deef2f6ed89c55372ac9b6afdd2
GitOrigin-RevId: bad76a902ccedef189ea31e816d6386b1458f143
Change-Id: I607a1a52e8a266df06f8b4a3b478873087fc7fe3