[roll] Roll fuchsia [system-updater] Use resolved update package hash when necessary
If the update package URL is unpinned, system-updater will now use the
hash from the resolved package when setting the retained package index
instead of just not adding the update package to the retained index.

This bug has not been an issue so far because in production the update
URL is always pinned and on eng builds (which use
system-update-checker which uses the unpinned
"fuchsia-pkg://fuchsia.com/update" URL) the package would be protected
by the dynamic index.

Upcoming changes for open package tracking will migrate system-updater
to a PackageResolver capability that no longer adds packages to the
dynamic index, regardless of the state of the retained index.

Original-Fixed: b/304607597
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929933
Original-Revision: f88abe2d2a0264e04ce5c60129f1215744946aa3
GitOrigin-RevId: 446b369b652295987cda15644971ddd2036e66b0
Change-Id: Ibf02957467498ee234e54e538883d127bddd5673