[roll] Roll fuchsia [device-enumeration-test] Ignore xhci root hub on vim3
We've seen many (12+) devices that often fail the device enumeration
tests because the XHCI hub doesn't enumerate.  But we don't actually
make use of the XHCI bus in any way, and we'd rather not have to either
tolerate this tests constantly flaking nor do we want to have to pull
the vim3s from the hardware fleet.  So just ignore when this device
fails to enumerate instead for now, and we can revisit if we start doing
anything with the XHCI bus.

Original-Bug: b/296738636
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930552
Original-Revision: 26efe9279169ee82c8089b2c5dda0b55d21ffcbc
GitOrigin-RevId: 8d8c7f9783eb2712c9671030b4dfe98ab8c6519e
Change-Id: Ie92737427ccdf2ba1cf91abd1fb218b99a5a5659