[roll] Roll fuchsia [ktrace] Fix Streaming Mode in Ktrace
We copy all the data from the kernel buffer to the trace buffer in a
loop in an async loop task which blocks the buffer saving mechanism that
also runs on the async loop. Allowing the copy to release the loop and
reschedule itself allows the streaming mechanism to work properly.

Original-Fixed: b/304349111
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929674
Original-Revision: b1208227d3580f203c6a04e4115f9de83059bdb1
GitOrigin-RevId: f252eea40c01f508118776faa82d370b7f50c1a6
Change-Id: I736c0ba16035e857e94ee8ec5a8b76cb8a9fb118