[roll] Roll fuchsia [bedrock] Router routing prototype
This CL introduces a lego block for asynchronous/lazy routing called
Router. A Router takes a request to obtain a capability and may either
resolve the request, transform it, or forward to another Router. So it
is similar to fuchsia.io/Open or CapabilityProvider in spirit, but
adapted to support both pipelined and un-peered capabilities.

This CL shows that component outputs, program outputs, dictionaries, and
availability transformations can all be modeled as a Router. The precise
implementation inside the Router is of course specific to the capability
type, but you can plug these together to achieve flexible routing.

Router itself is also a capability so you can put it in a dictionary
etc.

Original-Fixed: b/299994258

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/920254
Original-Revision: 1719dae51d91b4a1eb937596f827c90a50c3820c
GitOrigin-RevId: 6d65d21655d293efec0518f22a72afc292ec052a
Change-Id: I5f9ed64ca4aa8f1d56ef63372b73f87b6fce2cec