[roll] Roll fuchsia [kernel][riscv] add icache flush logic
In the usual places where the i and d cache need to be synchronized add
fence.i and the SBI call to sync on all of the other cores. QEMU does
not emulate this but is necessary on real hardware.

fence.i is present in basic privileged spec, so no extensions are
required.

Original-Bug: 124333

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/914499
Original-Revision: 41b305170fb7e695eee33bb035621bb299b68206
GitOrigin-RevId: 792666f3f397c802f1dc453749805b268fdfff46
Change-Id: I9c59671dd0a75d5b54ee6c170b32bbc4a786137f