[roll] Roll fuchsia [kernel][debug] Move mdscr from iframe to arch_thread
Remove the mdscr from the iframe_t on arm and move it into arch_thread.
Keeping it in the iframe_t makes it difficult to manage the value across
restricted/normal-mode thread transitions.

As implemented this will share the mdscr_el1 value for the thread in
both normal and restricted mode. We can mitigate this by managing this
value as part of Restricted/Normal state in r-mode, however this CL
doesn't implement that.

To test this change, I wrote a new zxdb e2e test that sets a hardware
breakpoint on `hello_starnix`, and enabled running this test on vim3.

Original-Bug: 131121
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895026
Original-Revision: 11a385394ae5cd9ac96df16318d22c4370ce0799
GitOrigin-RevId: 3f9264580b12a0b490f62cda5d54a2acd11d5549
Change-Id: I86cff2f56b3d5be83d9588c4670db7e9a8c79a03