[roll] Roll infra/recipes [cipd-util] Add option to fail build on existing package
The rust toolchain had a misconfiguration in their builder setup which
resulted in a race condition between two builders as to which one would
upload a package when only one of them was intended to. This change adds
a parameter to cipd_util.upload_package to fail the build on a package
already existing to make the builder fail loudly if this happens going
forward.

Original-Bug: 134695
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/927701
Original-Revision: 5417a148b3efca932777d29e64416dcde53e24e5
GitOrigin-RevId: e5ddec3d105a77e87a1a80f147e48af6cb0fe271
Change-Id: Iec74ce03e0fe988a479df289f292c9825789a46c