[roll] Roll infra/recipes [cipd] Always output deferred steps
The code as written will output deferred steps only in the case of a
changed CIPD package that was uploaded. This is the correct eventual end
state, as recipe_wrapper will want to modify the uploaded packages. At
this point in testing, rw won't do that, nor is it set to upload results
to SCILo.

In order to see things working, move the code such that deferred steps
are always written out.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/929899
Original-Revision: 3d3e7305c40a9e1dd4a5621256c9a5a5771d62ad
GitOrigin-RevId: 2aa6ebf5449c644b09a97674c57fa4d9d88ffc51
Change-Id: I49e1a54517b4300182920eaabf266b27dc2f7e67