[roll] Roll fuchsia [bazel-sdk] allow fuchsia_package to set its platform
This change allows a fuchsia_package to specify what platform it will be
built for. This lets us have platform specific packages. There is more
work that needs to be done here but we are introducing this change
gradually to allow adoption.

This change does not add target_compatible_with flags to the package.
This will be done later when we clean up our transitions and tests.

BUG: b/303448155
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930560
Original-Revision: 5dd3bff7ac54d428c1dacc346f6f774cac40e0e8
GitOrigin-RevId: d935a9f6fb83ee82c6d98f581a0efae75e69c068
Change-Id: I38290797d70584b50f451b07dbfe72fa750709b2