[roll] Roll fuchsia [netstack3] Remove non-FIDL addresses from core
Currently, fuchsia.net.interfaces.admin/Control.RemoveAddress can only
remove addresses that are present in the bindings HashMap, i.e.
addresses that have a corresponding FIDL worker / AddressStateProvider.
When an address to be removed is not present in that HashMap, we should
fall back to removing it from core (as it is a SLAAC or loopback
address).

Original-Bug: 135102
Multiply: interfaces-admin
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930232
Original-Revision: 30a003239849ab6cb260f855bb5d1fc648d8b68d
GitOrigin-RevId: 991235fbc5ea600a4eb8e2f832d90762c5ab9bc6
Change-Id: I0a6903e946d159c60ab6a792fcc6cbe8f921fb91