[roll] Roll fuchsia [kernel][phys] Make page table allocation internal to AddressSpace
Previously, an allocator was passed to AddressSpace by the user on
initialization. This was to account for a warty x86-64 case in which
allocation needed to be restricted to a portion of .bss in order to
bootstrap a new virtual address space; otherwise, a simple
memalloc::Pool-based allocator was used.

However, we eventually want to introduce a global AddressSpace instance
that lives past this initial set-up and to later use this to make
permanent kernel mappings. This means we need to ensure that the pool
tracks the page tables the back all such mappings. Since this
determination is already statically known within AddressSpace's
definition, it is cleanest to keep the allocation details internal to
the class. In the above x86 case, we do this by freeing the .bss region
for allocation and then introducing an AddressSpace knob for temporarily
clamping page table allocation to that region.

Original-Bug: 91187
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927136
Original-Revision: cd3db7a09f9c1712174a5656441d1b8edda0846c
GitOrigin-RevId: 86c1a0915b8cf2313179d871982e57649295cb0d
Change-Id: Ib52b310de537050200e2c59d9bef1d9429c235de