[roll] Roll fuchsia [fshost] add fxblob support to wipe_storage
This does not add support to the recovery workflows which call
wipe_storage. An additional argument is added to wipe_storage which lets
the caller connect to the new blob creator protocol, which when passed
in on a system configured for fxblob will be connected.

This required some changes to the disk builder to make sure fxblob could
be written on a gpt partition, which is a requirement for wipe_storage.

Original-Fixed: b/293970561
Test: fx test fshost_integration_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929352
Original-Revision: 54f0431a6a8bc21db8b864d9612d253d3da7555a
GitOrigin-RevId: 65ff4d824e03525e5ee0cc130c1a61c776c89e76
Change-Id: I611c6705edd9eda7c2c963e8d88302084ebbf9dc