[roll] Roll fuchsia [system-updater] Use the OTA specific PackageResolver capability
pkg-resolver's OTA specific resolution capability does not add
packages to the dynamic index, regardless of the state of the retained
index.

With this change, non-OTA package resolution is the only client of the
dynamic index, so we can replace it with open package tracking.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930032
Original-Revision: 12384e20632771c94178f00733d6f57c5f8c9885
GitOrigin-RevId: d3b372699536f120664b8cfd858a0cabd78f5761
Change-Id: Iecb1bb25d8e7ca79807d18572477e49364e0c8a9