[roll] Roll fuchsia [uart][ns8250] Check the LSR.THRE bit instead of TEMT for determining if TX is ready
LSR.THRE (Transmitter Holding Register Empty) tracks that the FIFO
is empty and thus a full FIFO sized payload of characters can be
dumped into it. More importantly it follows precisely the definition
of the THRE interrupt so can be used safely to determine when to enable
or disable the transmit interrupt.

LSR.TEMT (Transmitter Empty) tracks that in addition to the FIFO being
empty, the shift register holding the current character being shifted is
empty.

This seems like a generally more strict signal for TxReady() to follow,
but in the driver this could cause the THRE interrupt to be armed while
TxReady() is still saying it's not ready, causing an interrupt storm
while the last character is being shifted out.

Found on a real device with a dw8250 core. Emulators will probably never
simulate at this level since characters will likely instantly dissapear
from the fifo, or at least never spend time in a shift register that
isn't present.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929686
Original-Revision: bc4e71c8ff6e25abab6d4e2ff2da5ef27fd5967e
GitOrigin-RevId: 5ba144ac95fbfc95d47330594ec1486dd9e71dc5
Change-Id: Ic38996eb8b2e454ecdb468b873f1daf4991e5098