[roll] Roll fuchsia [debugger] Remove llvm::DWARFContext from indexing.
This uses the context on the DWARF binary. The index previously used a
separate context for some operations, but other operations used the one
on the DwarfBinary object. This standardizes on the DwarfBinary's one.

If we need to clear cached state after indexing, we should have a
duplicate DwarfBinary (as covered by the TODO in module_symbols_impl.cc
when the indexing code is called).

There should be no behavior change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/931492
Original-Revision: d6b31d5406f43199efdd4b44e7bd4a7b47f7b350
GitOrigin-RevId: ede71475a08d4ed3eaeda25c5396d9fc4ff39174
Change-Id: I20e925e268540a289822753c54f57a18a789a7a3