[roll] Roll fuchsia [netstack3] Avoid panic in UDP receive_icmp_error
This CL addresses a TODO to remove the `assert_is_ip_socket` function.
This was a runtime assert verifying that a dual-stack IPv6 socket never
receives an ICMPv4 error. When this assert was added, dual-stack IPv6
sockets couldn't yet be created; however, that is no longer the case.

At the moment it's not obvious how ICMPv4 errors should be handled by
IPv6 sockets, so I'm leaving that as a TODO for now.

Original-Bug: 135202
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/931372
Original-Revision: c8608b3f05dd62cb1787892da5e6741651c40cab
GitOrigin-RevId: 1958ac41882e7be35171fc9d9cfbabd3071aec16
Change-Id: Ida34b0b1d2628500ba149fbb70827eaee65fda94