[roll] Roll infra/recipes [luci_config] Migrate to new LUCI Config service
The old luci-config.appspot.com service is being deprecated in favor of
a rewritten version that uses a gRPC API. See bug for more context.

The easiest way to make RPC calls to a LUCI service from a recipe is to
use the `prpc` tool, which is guaranteed to be on $PATH for Buildbucket
builds.

Configs over 30MB will not have their content included in the RPC
response and must instead be fetched via signed URL. I didn't
implement support for signed URL fetching yet because we don't have any
configs over 30MB, so it would be difficult to test that logic and
there's no need for it.

Original-Bug: 135332
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/932352
Original-Revision: 8d60cce080e53f210260ba3e4141e190e0d0c646
GitOrigin-RevId: c8692f781e0b7b52d5520a218e6b886955382221
Change-Id: Ibe56613a9e4b97be6bce410a158f608596a99308