[roll] Roll fuchsia [doc-checker] Enabling validating reference links
This enables checking markdown reference links which use [<label>]
to reference pages.

Strictly parsing markdown also allows reference links to use
the contents between [] to be interpreted as a reference link label.
As a result, usages of [] that do not appear to be intended as
reference links are listed as warnings. Ideally, usage of [] should
be escaped with a single backslash (\).

Original-Bug: b/289246762
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930294
Original-Revision: ff3d188da655308446c272c2ea34cdc12ca59960
GitOrigin-RevId: 646786ba6fea8698578ac46c55f33f2909d0e0fa
Change-Id: I10df5b0ac95d2244a21dbd91c49cfbed50a2cf2a