[roll] Roll fuchsia [bedrock] Move open_flags from Open to Directory
This is a refinement to what goes into an Open capability.

In order to mint a capability that can e.g. open into an outgoing
directory of a program, we'll want to turn a fuchsia.io/Directory handle
into an Open. However, Open today has an extra baggage. It requires the
open_flags argument. There's no good value for this argument when
creating an Open capability to open within an outgoing directory, since
we don't know what are the rights on the outgoing directory.

The open_flags argument is required in order to remote the Open into a
Directory FIDL, because each fuchsia.io/Directory connection is
associated with some rights.

In this CL, the remoted protocol for Open is instead
fuchsia.io/Openable. An Openable FIDL doesn't have directory rights. It
only lets you open more items relative to it. This should be a better
fit.

The flags are added when converting an Open into a Directory capability,
at which point the remoted protocol changes from Openable to Directory
as well.

Multiply: attributed-logging-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930855
Original-Revision: 3b953f41c57af8ac6db2fe4166ef276805389199
GitOrigin-RevId: e3af26a0a164042e0c05d1e841e331c7fc4d32ae
Change-Id: Ibfd673f170d4c253c08d02ab8e468587101e4f17