[roll] Roll fuchsia [pkg-cache] Retained index no longer blocks dynamic index
The retained index used to prevent packages from being added to the
dynamic index, so that packages downloaded during OTA would not be
protected by the dynamic index (so that system-updater could GC them
if necessary). system-updater now has an OTA exclusive package
resolver capability that never adds packages to the dynamic index, so
this blocking behavior is no longer necessary.

This CL removes the blocking behavior so that, on builds where
ephemeral and OTA resolves are intermingled, the ephemerally resolved
packages will be protected as long as they are in use (once open
package tracking is enabled).

Removing the blocking behavior also makes the behavior of the system
easier to predict, as there is one fewer dependence on hard-to-observe
state.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/932284
Original-Revision: 606a0afb8e8d6a49cbb2f65cdff03734d31487b4
GitOrigin-RevId: a22ffcfcb336b795d4b5c45cd6d5e3c877e3c878
Change-Id: Ie13f249882468bfbae116ec1aef0b1f02dfb0794