[roll] Roll fuchsia [display][api-types-cpp] Use back porch in DisplayTiming.
Previously DisplayTiming uses active lines, front porch, sync
width and blanking to express the timing. This matches the
definition in EDID / DisplayID / CTA timing definitions.

However, one drawback is that the blanking is an aggregate
value and there's an implicit precondition that the blanking
must >= front porch + sync width.

This change replaces the "blanking" fields in DisplayTiming
with "back porch" fields. The latter is atomic, so the timing
is valid as long as the back porch is in [0, kMaxTimingValue]
and we don't need any extra comparison.

For convenience, this change also adds horizontal_blank_px(),
vertical_blank_lines(), horizontal_total_px() and
vertical_total_lines() for drivers to access the aggregate
values.

Test: Updated all the tests in display-timing-test
      Added DisplayTiming.AggregateHelpers test case.
Original-Bug: 128286
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930315
Original-Revision: 14bdb5402efa980bf04b37858aacffca05ff6d12
GitOrigin-RevId: 8c48af58047c0ff4a3709dea4039dff7e5cc18bb
Change-Id: Idd9e8a13997543fd0a59b4a1bfa2e8c5d9c5a590