[roll] Roll fuchsia [kernel][boot-shim] Refactor devicetree memory reservation processing
We've come across a bootloader that creates memory reservations in its
devicetree for the ramdisk and devicetree blob itself. This is not great
and contrary to the devicetree spec itself (as these should represent
memory that the kernel should not use), but we should work around this
quirk nonetheless.

Currently the translation of memory reservations to RESERVED ranges
is handled by the devicetree memory matcher. We pull this processing
out of the matcher for two reasons:
* This logic is unrelated to matching, as the memory reservations are
not stored within the encoded tree structure;
* We need to filter reservations that represent things like the ramdisk,
and the ramdisk is a product of matching.

This new routine has straightforward logic for ignoring intersections
of the reservations with a given set of ranges to exclude.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929326
Original-Revision: 2b369c3552fb133f4152f21df999377a65e5cd0c
GitOrigin-RevId: 793fb5357569065b7696e40e2b46d6abc589cd56
Change-Id: Idfd2c29f88f97de5d5d396a9f3cd29f42c105124