[roll] Roll fuchsia [driver][index] No return for AddCompositeNodeSpec
Makes a change to the driver index's AddCompositeNodeSpec FIDL protocol
to not return a matched driver for the added composite spec. This is
because the driver matching process is how the driver flows to the
driver manager, the matched driver here has no effect. Previously we
were returning a not_found error when a match did not happen, but this
error is not important for the driver manager so now we return
successfully even when no match happens. This is because the composite
has still been successfully added.

Original-Bug: 111076
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/925373
Original-Revision: df878bba94714e05af2c6a8c7e39964f7eda9552
GitOrigin-RevId: 5b4abd361f83e7392dc7b7bb867e541c807922bc
Change-Id: Idbaf929b4953ecfe191da869c054b6962fb96d16