[roll] Roll fuchsia [ffx][golden files] Print diff when mismatched
This adds displaying the unified diff between the candidate and
golden files when there is a difference.

This improves the experience of dealing with command line golden file
changes since the change causing the difference is displayed.

Since there are also some usages that compare binary files,
a flag `binary` is added to perform the filecmp based comparison
and not display the diff.

Original-Bug: b/303542110
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/932258
Original-Revision: baa8fa9a30ee54aa7a383e12fc7ed957408e0342
GitOrigin-RevId: 6b07f8a1c7e86063da461b8c1fe27c2019accc19
Change-Id: I088df9ccdd564e76d71f8e01c7c89d11ff23d0ae