[roll] Roll infra/recipes [platform_version_roller] Fixes.
The platform version roller made various references to the "current"
platform version, which really meant, the first in-development API
level found in `version_history.json`. Currently, there's _no_
in-development API level, so things broke.

On top of that, the "current platform version" isn't really a thing.
We've got milestone branches and API levels, and this roller compares
the two and sees if it should trigger creation of a new API level. So
I've changed the roller to reflect that framing.

Original-Bug: b/305982333

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/933912
Original-Revision: b81e27b10464c42d9978e948bf8ad23fafc1016d
GitOrigin-RevId: eeb34acc6464912c8a035697eb63af6e07de2f11
Change-Id: I8550c673f4369382cbdcea965c1000efdc324820