[roll] Roll fuchsia [component_manager][config] Remove duplicates on vector merges
It's unclear how Component Framework would respond to duplicate
strings, so its simpler to remove them.

The deduping process is updated to sort and then dedup instead
of using a hashmap. The main benefit of this is that the
returned vector is sorted, which gives it a deterministic
output, so it is easier to test and compare.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/931072
Original-Revision: 4d50d0350de07bfd36cd425b7c82889e29a88e80
GitOrigin-RevId: a42447d5d7adbf854f576b5a44d5eed7411aaabf
Change-Id: I930583854585110f836dd1511373d782f5033f27