[roll] Roll fuchsia [sandbox] Implement duck typing on Handle for cloning and TryInto<Open>
Implement try_clone and TryInto<Open> on Handle:

* try_clone clones channels that speak fuchsia.unknown.Cloneable
* TryInto<Open> converts channels that speak fuchsia.io.Openable

When a handle is inserted into a Dict, it's stored as a Handle type
as before - there is no attempt to determine its "original" type.

Then, when a client needs the handle to do a particular behavior,
like cloning, or converting to a different type, the Handle type
does this opportunistically.

* Add a test from fxrev.dev/886692
* Sender now implements Openable and Cloneable so Dicts that
  contain senders as Handle types can be converted to a namespace.

Original-Bug: b/303307609
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929300
Original-Revision: 1a5beb9d9041307b8d0a7d15d147268f9fa21a22
GitOrigin-RevId: 37a64463b6728afed6bc9cf50c2205260295ac34
Change-Id: Ia27a465411a95c2be2a760de8e673b9b1572dbed