[roll] Roll fuchsia [ffx] use env variables for target default
Instead of having no target default, set the default to be a fallback
list of environment variables:
  FUCHSIA_DEVICE_ADDR
  FUCHSIA_NODENAME
(But of course these will be overridden as always by "ffx -t" or by
"ffx target default set".)

This should provide more consistent behavior across different sets
of users (including automated systems which often set the target
via one of the above environment variables).

Note that semantics of ADDR vs NODENAME is different: the former
is an address, the later is a name. But when used in ffx, they
are interchangeable since a target query looks at both the address
and the name to find a target.

Original-Bug: b/297896809
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/932575
Original-Revision: 20c02691216da6a0aca76541c1643b21ba4311aa
GitOrigin-RevId: c76807cbc8e764dc26f0f2d6d13267fe0cbce8df
Change-Id: Iac23ac38ccfcea42a456c3fa9f051fce27310064