[roll] Roll fuchsia [elfldltl][ld] Static TLS layout
This adds a toolkit class elfldltl::TlsLayout for doing the
static TLS layout calculations in accordance with the psABI.
This uses new a elfldltl::TlsTraits template struct specialized
for each machine to describe the relevant psABI constants.

The startup dynamic linker uses this to perform static TLS layout
for each module with a PT_TLS.  The calculated totals and the
per-module offsets are exported in the passive ABI.

So far this is only tested by examining the passive ABI.
There is not yet any support for the TLS relocations.

Original-Bug: 128502
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929896
Original-Revision: 5de86e3ff35476efa5a64d75e7b8de2502c509fa
GitOrigin-RevId: 47930baf06bd80ae6a927ae7c4ee1cd9005b6d1d
Change-Id: Ie7bd905c7b27fe7c64144a6f93a49efe6e12bf0d