[roll] Roll fuchsia [ffx][tc] Implement TargetUpdate to describe mutations
This struct captures the various ways targets are updated due to
discovery and RCS. Connection states are represented in TargetUpdate in
a simpler way.

NOTE: For backwards compatibility there are no changes to the target struct.

Discovery and RCS changes will be in a follow-up CL.

Original-Bug: b/301988834
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927334
Original-Revision: 4ebada8de07536698c58963de547b93930e8fad6
GitOrigin-RevId: 0e0fab59c81240708042371fbf4e3f3bc06f1311
Change-Id: I80098253a4136fa0a3e33fd781be68bd245327ef