[roll] Roll fuchsia [fdf] Check ancestor nodes when deciding node collection
Set driver framework nodes' collections to be equal to or higher ranked
than all of their ancestors' collections instead of just their parents'
collection. This is to avoid node components from having cyclic
dependencies on each other.

Sometimes a node's parents may not have a collection because the parent
nodes don't have drivers bound to them that require a component. In that
case, a node would need to check their parents' parents in order to
determine the which collection to be put in.

Original-Bug: 134719
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934136
Original-Revision: 599257c8829da9d7833134ad4187095f86c5909d
GitOrigin-RevId: 3db2ca9d2d938cea263f30840641822fd9faafd6
Change-Id: If8e83d6b1b84cf8c512bacbf7ce75998c5d0478a