[roll] Roll scudo [scudo] Add ConditionVariable in SizeClassAllocator64 (#69031)
This may improve the waiting of `Region->MMLock` while trying to refill
the freelist. Instead of always waiting on the completion of
`populateFreeListAndPopBatch()` or `releaseToOSMaybe()`, `pushBlocks()`
also refills the freelist. This increases the chance of earlier return
from `popBatches()`.

The support of condition variable hasn't been done for all platforms.
Therefore, add another `popBatchWithCV()` and it can be configured in
the allocator configuration by setting `Primary::UseConditionVariable`
and the desired `ConditionVariableT`.

Reviewed By: cferris

Differential Revision: https://reviews.llvm.org/D156146

GitOrigin-RevId: 19bdc182d26e7818c39ef59d28b8c09455691d80
Original-Revision: 346790be383eed685ea7752733d7ed6d3f93f833
Change-Id: Ic858972bc621dd50c7bb3691a34ea46bc07228c1