[roll] Roll fuchsia [kernel][x86][reset] refactor the 0xcf9 reset mechanism to allow AMD cores to use it
IO port 0xcf9 is actually a fairly standard way to reset PCs with PCI
busses, and according to the 15h and above PPRs it's also implemented on
AMD cores from Bulldozer and above (and probably much earlier as well).

Refactor the previously intel specific mechanism to be a bit more
generic and include the AMD cores as well.

This should allow for rebooting with a reason on anything reasonably
modern AMD wise.

Original-Bug: 135095
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/931529
Original-Revision: 6144d50a7e10a5fce57a470e8daf93719024cb42
GitOrigin-RevId: a6532635c9efb0aa9cf9cd4c0c8885d750aeca04
Change-Id: I3b5ead15b22d7fe298420d81376f1ce57837f072