[roll] Roll fuchsia [routing] Create NoopVisitor
Make a generic NoopVisitor and then type alias
it to make it easier to use.

This removes a macro, and it also makes it
possible to use a generic NoopVisitor if
we don't know the type ahead of time.

Original-Bug: b/305948828

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934593
Original-Revision: 969e202136ae02b8e583dfd40acf535bda1f7fd2
GitOrigin-RevId: 9af2520255344129efae1ae86662cf4a803290ae
Change-Id: I6e9e274fcaefecc2aa9fea2748780444a401f57a