[roll] Roll fuchsia [routing] Remove mapper from Aggregates
The mapper is cloned to be placed in the aggregate classes.
This means the mapper can't actually be used by scrunity
because the routes are being added to a new mapper.

Remove the mappers from these classes so we can remove
Clone from the Mapper class.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934912
Original-Revision: e447921ad2c8448ffd3482c197d085c68fd12535
GitOrigin-RevId: b23223ca7ed8e57de317d995969fa4e8cc599378
Change-Id: Ia593f0ab4f37ed2474c631293e9ee624c2cd4abe